+++ paper/paper.bib
@@ -2,6 +2,7 @@
   title={A solar C/O and sub-solar metallicity in a hot Jupiter atmosphere},
   author={Line, Michael R and Brogi, Matteo and Bean, Jacob L and Gandhi, Siddharth and Zalesky, Joseph and Parmentier, Vivien and Smith, Peter and Mace, Gregory N and Mansfield, Megan and Kempton, Eliza M-R and others},
   journal={Nature},
+  doi = {10.1038/s41586-021-03912-6},
   volume={598},
   number={7882},
   pages={580--584},
@@ -13,17 +14,28 @@
   author={Mansfield, Megan and Wiser, Lindsey and Stevenson, Kevin B and Smith, Peter and Line, Michael R and Bean, Jacob L and Fortney, Jonathan J and Parmentier, Vivien and Kempton, Eliza M-R and Arcangeli, Jacob and others},
   journal={The Astronomical Journal},
   volume={163},
+   doi = {10.3847/1538-3881/ac658f},
   number={6},
   pages={261},
   year={2022},
   publisher={IOP Publishing}
 }
-@article{august:2023,
-  title={Confirmation of sub-solar metallicity for WASP-77Ab from JWST thermal emission spectroscopy},
-  author={August, Prune C and Bean, Jacob L and Zhang, Michael and Lunine, Jonathan and Xue, Qiao and Line, Michael and Smith, Peter},
-  journal={arXiv preprint arXiv:2305.07753},
-  year={2023}
-}
+@ARTICLE{august:2023,
+       author = {{August}, Prune C. and {Bean}, Jacob L. and {Zhang}, Michael and {Lunine}, Jonathan and {Xue}, Qiao and {Line}, Michael and {Smith}, Peter C.~B.},
+        title = "{Confirmation of Subsolar Metallicity for WASP-77Ab from JWST Thermal Emission Spectroscopy}",
+      journal = {\apjl},
+     keywords = {Exoplanets, Exoplanet atmospheres, James Webb Space Telescope, 498, 487, 2291, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2023,
+        month = aug,
+       volume = {953},
+       number = {2},
+          eid = {L24},
+        pages = {L24},
+          doi = {10.3847/2041-8213/ace828},
+archivePrefix = {arXiv},
+       eprint = {2305.07753},
+ primaryClass = {astro-ph.EP},
+ }
 @article{buchner:2014,
   title={X-ray spectral modelling of the AGN obscuring region in the CDFS: Bayesian model selection and catalogue},
   author={Buchner, J and Georgakakis, A and Nandra, K and Hsu, L and Rangel, C and Brightman, M and Merloni, A and Salvato, M and Donley, J and Kocevski, D},
@@ -31,6 +43,7 @@
   volume={564},
   pages={A125},
   year={2014},
+   doi = {10.1051/0004-6361/201322971},
   publisher={EDP Sciences}
 }
 @article{line:2013,
@@ -39,6 +52,7 @@
   journal={The Astrophysical Journal},
   volume={775},
   number={2},
+  doi = {10.1088/0004-637X/775/2/137},
   pages={137},
   year={2013},
   publisher={IOP Publishing}
@@ -49,6 +63,7 @@
   journal={Philosophical transactions of the royal society A: Mathematical, Physical and Engineering Sciences},
   volume={374},
   number={2065},
+  doi = {10.1098/rsta.2015.0202},
   pages={20150202},
   year={2016},
   publisher={The Royal Society Publishing}
@@ -58,6 +73,7 @@
[…]
