### 2023-11-21 9723b70
generalize CIA to work on different filetypes
