+++ src/cortecs/opac/io.py
@@ -9,6 +9,7 @@ import pickle
 
 import h5py
 import numpy as np
+import pandas as pd
 from tqdm import tqdm
 
 AMU = 1.6605390666e-24  # atomic mass unit in cgs. From astropy!
@@ -202,35 +203,57 @@ class loader_exotransmit_cia(loader_base):
     loads in opacity data that are used with the PLATON code's collision-induced absorption..
     """
 
-    def load(self, cross_sec_filename, T_filename, wl_filename, species_name):
+    def load(self, cross_sec_filename, verbose=False):
         """
-        loads in opacity data that's built for PLATON. To be passed on to Opac object.
+        loads in opacity data that's built for exo-transmit. To be passed on to Opac object.
 
-        The temperature grid, pressure grid, and wavelength grid are saved as separate files for PLATON.
 
         Inputs
         ------
-        filename : str
-            name of file to load
-        cross_sec_filename : str
-            name of cross section file
-        T_filename : str
-            name of temperature file
-        wl_filename : str
-            name of wavelength file
-        species_name : tuples
-            name of two colliding species. E.g., ('H2', 'CH4'). todo: all at once?
+            :CIA_file: (str) path to CIA file. e.g., 'opacCIA/opacCIA.dat'
+        Outputs
+        -------
+            :df: (pd.DataFrame) dataframe with the CIA data.
         """
-        # todo: check wl units. They're in meters here.
-        # temperatures are in K.
-        # pressures are in Pa, I believe.
-        wl = np.load(wl_filename)
-        T = np.load(T_filename)
-        data = pickle.load(
-            open(cross_sec_filename, "rb"), encoding="latin1"
-        )  # packaged as T x P x wl. todo: check packing
-        cross_section = data[species_name]
-        return wl, T, cross_section
+
+        f = open(cross_sec_filename)
+        f1 = f.readlines()
+        f.close()
+
+        temperatures = []
+        wavelengths = []
+
+        species_dict = get_empty_species_dict(cross_sec_filename, verbose=verbose)
+
+        # read through all lines in the CIA file
+        for line in tqdm(f1[1:], desc="Reading CIA file"):
+            if not line or line == "\n":
+                continue  # don't want it to break!
+            if len(line.split()) == 1 and line != "\n":  # this is a new temperature
+                temperature = eval(line[:-1])
+                continue  # nothing else on this line
+            values = [eval(value) for value in line.split()]
+            temperatures += [temperature]
+            wavelengths += [values[0]]
+
+            for species_ind, species in enumerate(species_dict.keys()):
+                species_dict[species] += [values[species_ind + 1]]
+
+        # hm. how to do this with different numbers of species?
+        df = pd.DataFrame(
+            {
+                "temp": temperatures,
+                "wav": wavelengths,
+            }
+        )
+        for species in species_dict.keys():
+            df[species] = species_dict[species]
+
+        columns = list(df.columns)
+        columns.remove("temp")
+        columns.remove("wav")
+
+        return df.wav.values, df.temp.values, df[columns]
 
 
 class loader_exotransmit(loader_base):
@@ -309,4 +332,181 @@ class loader_exotransmit(loader_base):
         return wl, T, P, opacities
 
 
+# todo: put these in the opac class?
+def get_empty_species_dict(CIA_file, verbose=False):
+    """
+    returns a species dictioanry given a CIA file
+    :param CIA_file:
+    :param verbose:
[…]
