+++ src/cortecs/opac/opac.py
@@ -3,6 +3,8 @@ this file holds the class describing opacity data. hmm...maybe I want the loader
 
 author: @arjunsavel
 """
+import numpy as np
+
 from cortecs.opac.io import *
 
 
@@ -72,6 +74,11 @@ class Opac(object):
             )
         return self.method_dict[loader_name]
 
+    # todo: implement the copy and deepcopy methods.
+    def copy(self):
+        # need a new opac object with all the same attributes. so let's make sure we don't load on input.
+        raise NotImplementedError
+
 
 class Opac_cia(Opac):
     """
@@ -79,7 +86,6 @@ class Opac_cia(Opac):
 
     Everything's already been loaded into memory, so this is just a wrapper for the data.
 
-    todo: add exotransmit CIA
     """
 
     method_dict = {
@@ -87,9 +93,10 @@ class Opac_cia(Opac):
         "exotransmit_cia": loader_exotransmit_cia,
     }
 
-    def __init__(self, filename, loader="platon_cia"):
+    def __init__(self, filename, loader="platon_cia", view="default"):
         """
         wraps around the loaders.
+        NOTE: there's a mixed type for self.cross_section. can be a dataframe if there are a number of species.
 
         Parameters
         ----------
@@ -97,16 +104,54 @@ class Opac_cia(Opac):
             name of file to load
         loader : str
             name of loader to use. default is chimera.
+        view : str
+            if 'full_frame' is selected, the cross_section object also includes temperature and wavelength columns.
 
         Returns
         -------
         nothing
         """
         self.filename = filename
-        load_obj = loader_base()
-        self.wl, self.T, self.cross_section = load_obj.load(filename, loader=loader)
+        self.load_obj = self.method_dict[loader]()
+        self.wl, self.T, self.cross_section = self.load_obj.load(filename)
+
         # at least for the exotransmit case, we have...wl x temp.
+        if view == "full_frame":
+            self.cross_section["temp"] = self.T
+            self.cross_section["wav"] = self.wl
 
         self.n_wav, self.n_t = len(self.wl), len(self.T)
+
+        return
+
+    def join_cross_section(self, opac):
+        """
+        joins another opacity's cross-section data to this one.
+        :param opac:
+        :return:
+        """
+
+        def replace_zeros(species):
+            mask = self.cross_section[species] == 0.0
+            self.cross_section.loc[mask, species] = opac.cross_section.loc[
+                mask, species
+            ]
+
+        # throw errors if the grids don't match.
+        if np.all(opac.T != self.T):
+            raise ValueError("temperatures do not match!")
+        if np.all(opac.wl != self.wl):
+            raise ValueError("wavelengths do not match!")
+
+        # add the entirely new columns
+        other_columns = opac.cross_section.columns
+        for column in other_columns:
+            if column not in self.cross_section.columns:
+                self.cross_section[column] = opac.cross_section[column]
+
+        # now fill in any temperatures where the cross-section is zero here.
+        for species in self.cross_section.columns:
+            if species in other_columns:
+                replace_zeros(species)
 
         return
