+++ src/cortecs/opac/interpolate_cia.py
@@ -22,158 +22,9 @@ import pandas as pd
 from tqdm import tqdm
 
 from cortecs.opac.chunking import *
+from cortecs.opac.io import *
 
 ######################### Pt. 1: Interpolation ####################################
-
-
-def get_n_species(CIA_file, verbose=False):
-    """
-    Returns the number of species in a CIA file.
-    Inputs
-    -------
-        :CIA_file: (str) path to CIA file. e.g., 'opacCIA/opacCIA.dat'
-    Outputs
-    -------
-        :n_species: (int) number of species in the CIA file.
-    """
-
-    f = open(CIA_file)
-    f1 = f.readlines()
-    f.close()
-    # the third line fshould be the first normal one.
-
-    line = f1[3]
-    n_species = len(line.split()[1:])
-
-    if verbose:
-        print(f"Number of species in CIA file {CIA_file}: {n_species}")
-        if n_species == 8:
-            print(
-                "Species are likely: H-el, He-H, CH4-CH4, H2-He, H2-CH4, H2-H, H2-H2, CO2-CO2"
-            )
-        elif n_species == 14:
-            print(
-                "Species are likely:  H2-H2, H2-He, H2-H, H2-CH4, CH4-Ar, CH4-CH4, CO2-CO2, He-H, N2-CH4, N2-H2, N2-N2, O2-CO2, O2-N2, O2-O2)"
-            )
-    return n_species
-
-
-def get_empty_species_dict(CIA_file, verbose=False):
-    """
-    returns a species dictioanry given a CIA file
-    :param CIA_file:
-    :param verbose:
-    :return:
-    """
-    n_species = get_n_species(CIA_file, verbose=verbose)
-    if n_species == 8:
-        # todo: refactor. has to be a cleaner way to do this! infer the columns, etc.
-        Hels = []
-        HeHs = []
-        CH4CH4s = []
-        H2Hes = []
-        H2CH4s = []
-        H2Hs = []
-        H2H2s = []
-        CO2CO2s = []
-
-        # python > 3.6 has ordered dictionaries!
-        species_dict = {
-            "Hels": Hels,
-            "HeHs": HeHs,
-            "CH4CH4s": CH4CH4s,
-            "H2Hes": H2Hes,
-            "H2CH4s": H2CH4s,
-            "H2Hs": H2Hs,
-            "H2H2s": H2H2s,
-            "CO2CO2s": CO2CO2s,
-        }
-
-    elif n_species == 14:
-        H2H2s = []
-        H2Hes = []
-        H2Hs = []
-        H2CH4s = []
-        CH4Ar = []
-        CH4CH4s = []
-        CO2CO2s = []
-        HeHs = []
-        N2CH4s = []
-        N2H2s = []
-        N2N2s = []
-        O2CO2s = []
-        O2N2s = []
-        O2O2s = []
-        species_dict = {
-            "H2H2s": H2H2s,
-            "H2Hes": H2Hes,
-            "H2Hs": H2Hs,
-            "H2CH4s": H2CH4s,
-            "CH4Ar": CH4Ar,
-            "CH4CH4s": CH4CH4s,
-            "CO2CO2s": CO2CO2s,
-            "HeHs": HeHs,
-            "N2CH4s": N2CH4s,
-            "N2H2s": N2H2s,
-            "N2N2s": N2N2s,
-            "O2CO2s": O2CO2s,
-            "O2N2s": O2N2s,
-            "O2O2s": O2O2s,
-        }
-    else:
-        print("Number of species in CIA file not recognized. Check the file!")
-        species_dict = {}
-        species_keys = np.arange(n_species)
-        for species_key in species_keys:
-            species_dict[species_key] = []
[…]

[…]
