### 2023-11-22 ac0f203
add explicit CIA testing, interpolation testing
