+++ src/cortecs/tests/test_opac.py
@@ -4,4 +4,277 @@ tests that the opacity object behaves as expected.
 
 import unittest
 
-from cortecs.opac import *
+from cortecs.opac.opac import *
+import copy
+import os
+import numpy as np
+from cortecs.opac.interpolate_cia import *
+
+
+class TestOpac(unittest.TestCase):
+    """
+    tests the Opac object.
+    """
+
+    cia_filename = (
+        os.path.abspath(".") + "/src/cortecs/tests/test_output_cia_lotemp.dat"
+    )
+
+    def test_wrong_method(self):
+        """
+        If I pass a method that isn't sypported, an error should be raised.
+        :return:
+        """
+        Opac = ""  # don't need a whole method here
+        self.assertRaises(ValueError, Opac, "wrongmethod")
+
+    def test_cia_opac_instantiated(self):
+        """
+        does an error get thrown when I'm just trying to make a little opacCIA object?
+        :return:
+        """
+        opac_test = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        self.assertTrue(True)
+
+    def test_cia_opac_join(self):
+        """
+        does an error get thrown when I'm just trying to join two things?
+        :return:
+        """
+        opac_test2 = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test2.join_cross_section(opac_test)
+        self.assertTrue(True)
+
+    def test_cia_opac_join_no_add(self):
+        """
+        shouldn't add any new columns if the second one has fewer columns.
+        :return:
+        """
+        opac_test2 = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test2.join_cross_section(opac_test)
+        self.assertTrue(
+            len(opac_test2.cross_section.columns)
+            == len(opac_test.cross_section.columns)
+        )
+
+    def test_cia_opac_join_yes_add(self):
+        """
+        shouldn't ONLY add the new columns if the second one has more columns.
+        :return:
+        """
+        opac_test2 = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        initial_copy = copy.copy(opac_test2)
+        opac_test = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test.cross_section["other_col"] = opac_test.cross_section["temp"] * 13
+        opac_test2.join_cross_section(opac_test.cross_section)
+
+        self.assertTrue(
+            len(initial_copy.cross_section.columns) + 1
+            == len(opac_test2.cross_section.columns)
+        )
+
+    def test_cia_opac_diff_temp(self):
+        """
+        needs same temp grid for joining.
+        :return:
+        """
+        opac_test2 = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
+        opac_test.T = [23]
+        self.assertRaises(ValueError, opac_test2.join_cross_section, opac_test)
+
+    def test_cia_opac_diff_wl(self):
+        """
+        needs same wl grid
+        :return:
+        """
+        opac_test2 = Opac_cia(
+            self.cia_filename, loader="exotransmit_cia", view="full_frame"
+        )
[…]
