+++ src/cortecs/opac/interpolate_cia.py
@@ -18,7 +18,6 @@ import os
 from glob import glob
 
 import numpy as np
-import pandas as pd
 from tqdm import tqdm
 
 from cortecs.opac.chunking import *
