+++ src/cortecs/tests/test_opac.py
@@ -16,9 +16,7 @@ class TestOpac(unittest.TestCase):
     tests the Opac object.
     """
 
-    cia_filename = (
-        os.path.abspath(".") + "/src/cortecs/tests/test_output_cia_lotemp.dat"
-    )
+    cia_filename = os.path.abspath(".") + "/src/cortecs/tests/test_output_cia.dat"
 
     def test_wrong_method(self):
         """
@@ -137,9 +135,7 @@ class TestOpac(unittest.TestCase):
 
 
 class TestIO(unittest.TestCase):
-    cia_filename = (
-        os.path.abspath(".") + "/src/cortecs/tests/test_output_cia_lotemp.dat"
-    )
+    cia_filename = os.path.abspath(".") + "/src/cortecs/tests/test_output_cia.dat"
 
     def test_write_cia_same_cross_section(self):
         """
@@ -206,9 +202,7 @@ class TestInterpolateCIA(unittest.TestCase):
     cross_sec_filename = (
         os.path.abspath(".") + "/src/cortecs/tests/absorb_coeffs_C2H4.npy"
     )
-    cia_filename = (
-        os.path.abspath(".") + "/src/cortecs/tests/test_output_cia_lotemp.dat"
-    )
+    cia_filename = os.path.abspath(".") + "/src/cortecs/tests/test_output_cia.dat"
 
     def test_interpolate_cia_within_bounds(self):
         """
