### 2023-11-22 12fa024
allow platon loaders in opacCIA
