+++ src/cortecs/opac/interpolate_cia.py
@@ -50,7 +50,9 @@ def check_temp_grid(df, real_temperature_grid, CIA_file):
     return
 
 
-def interpolate_CIA(CIA_file, reference_file, outfile=None):
+def interpolate_CIA(
+    CIA_file, reference_file, outfile=None, loader="exotransmit", loader_kwargs={}
+):
     """
     Interpolates a CIA file to a higher resolution, using the wavelength grid
     of a reference file. Note: This function assumes that the CIA file has
@@ -72,7 +74,7 @@ def interpolate_CIA(CIA_file, reference_file, outfile=None):
         Creates a file with "hires" attached to the end of CIA_file that has been interpolated
         to higher resolution.
     """
-    reference_opac = Opac(reference_file, loader="exotransmit")
+    reference_opac = Opac(reference_file, loader=loader, **loader_kwargs)
 
     real_wavelength_grid = reference_opac.wl
     # need to put it on the right temperature grid, too!
