### 2023-11-22 0fcb3bf
make a deepcopy instead of a copy!
