### 2023-11-22 0d7a04a
test the string printing...both
