### 2023-11-22 15be352
try a differnt string comparison
