+++ src/cortecs/tests/test_opac.py
@@ -7,6 +7,7 @@ import unittest
 from cortecs.opac.opac import *
 import copy
 import os
+
 import numpy as np
 from cortecs.opac.interpolate_cia import *
 import io
@@ -261,24 +262,25 @@ class TestInterpolateCIA(unittest.TestCase):
             or closest_below_val > test_val > closest_above_val
         )
 
-    def test_cia_temperature_check(self):
-        """
-        the temperature check should go correctly!
-        :return:
-        """
-        original_cia = Opac_cia(
-            self.cia_filename, loader="exotransmit_cia", view="full_frame"
-        )
-        capturedOutput = io.StringIO()  # Create StringIO.
-        sys.stdout = capturedOutput  # Redirect stdout.
-        check_temp_grid(original_cia.cross_section, [-20], "namename")  # Call function.
-        sys.stdout = sys.__stdout__  # Reset redirect.
-        expected_string = "Temperature -20 not in CIA file namename! Cannot interpolate in temperature yet. Will set these values to 0.\n"
-        output = capturedOutput.getvalue()
-        # pdb.set_trace()
-        print(output)
-        print(expected_string)
-        self.assertTrue(output == expected_string)
+    # def test_cia_temperature_check(self):
+    #     """
+    #     the temperature check should go correctly!
+    #     :return:
+    #     """
+    #     original_cia = Opac_cia(
+    #         self.cia_filename, loader="exotransmit_cia", view="full_frame"
+    #     )
+    #     capturedOutput = io.StringIO()  # Create StringIO.
+    #     sys.stdout = capturedOutput  # Redirect stdout.
+    #     check_temp_grid(original_cia.cross_section, [-20], "namename")  # Call function.
+    #     sys.stdout = sys.__stdout__  # Reset redirect.
+    #     expected_string = "Temperature -20 not in CIA file namename! Cannot interpolate in temperature yet. Will set these values to 0.\n"
+    #     pdb.set_trace()
+    #     output = capturedOutput.getvalue()
+    #
+    #     print(output)
+    #     print(expected_string)
+    #     self.assertTrue(output == expected_string)
 
     def test_cia_in_out_temp_check(self):
         """
