### 2023-11-22 f171f39
print the stuff instead of a valueerror
