### 2023-12-01 #6104 @editorialbot
[REVIEW]: cortecs: A Python package for compressing opacities
**Submitting author:** @arjunsavel (<a href="http://orcid.org/0000-0002-2454-768X">Arjun B. Savel</a>)
**Repository:** https://github.com/arjunsavel/cortecs
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
