### 2024-01-10 648e32c
avoid divide by 0 warning
