### 2024-01-10 b417a6e
remove script boilerplate
