+++ src/cortecs/fit/fit_pca.py
@@ -128,45 +128,3 @@ def prep_pca(cross_section, wav_ind=-1, nc=2):
 
     xMat, fStd, fNorm, beta, s, vh, u = do_pca(single_pres_single_temp, nc)
     return xMat
-
-
-if __name__ == "__main__":
-    # redo for the PCA. this is the big one!!
-    species_inds = {
-        "C2H2": 200,
-        "CO_hitemp": 169565,
-        "HCN": 200,
-        "NH3": 200,
-        "H2S": 200,
-        "H2O": 200,
-        "CH4": 200,
-        "H2H2": 200,
-        "H2He": 200,
-        "13CO": 169565,
-    }
-    with warnings.catch_warnings():
-        species_means_dict = {}
-        for species in ["CO_hitemp"]:
-            # '13CO', 'HCN', 'NH3', 'H2S', 'H2O', 'CH4',  'H2H2', 'H2He',  'C2H2',]:
-
-            full_opac = species_dict[species]  # todo: fix
-            wav_ind = species_inds[species]
-            species_means = []
-
-            single_pres_single_temp = full_opac[:, :, wav_ind]
-
-            xMat, fStd, fNorm, beta, s, vh, u = do_pca(single_pres_single_temp, 2)
-
-            betas = []
-
-            for i in tqdm(range(len(wno)), desc=species):
-                arr = full_opac[:, :, i]
-                _, beta = fit_mlr(arr, xMat)
-                diff = (arr - np.dot(xMat, beta).T) / arr
-                species_means += [np.median(np.abs(diff))]
-                betas += [beta]
-
-            species_means_dict[species] = np.array(species_means)
-
-#         np.save(f'{species}_vectors.npy', xMat)
-#         np.save(f'{species}_pca_coeffs.npy', np.array(betas))
