### 2024-01-10 abe8d6f
fix figure caption
