### 2024-01-10 2bfba52
trim PCA fitting
