### 2024-01-13 #6104 @pscicluna
I apologise for the delay, but here is my review!

### Summary:
The authors present an interesting and widely applicable software package. They are targeting its usage for dimensionality reduction of large opacity datasets for e.g. molecular gas, but I expect it could be applied much more widely (perhaps with small tweaks). For example, I expect that `cortecs` could also be useful for interpolation of opacities, including prediction uncertainty.

Similar to @doriannblain, I also find it a little difficult to see how `cortecs` would be integrated into an analysis pipeline. I imagine the idea is to start your pipeline by generating the compressed opacities, then (if necessary) distributing the evaluators to all the compute devices, then the sampler can run on all devices to get the posterior you’re after? If that’s the case, it would be a good idea to provide an example workflow, either in the docs or by describing it further in the paper.

I also find it hard to tell when `cortecs` can be deployed to maximum benefit, and when it should be avoided. For example, can it only be used for molecular gas opacities, or could it be used to compress atomic opacities, or even temperature-dependent dust opacities? How large and complex does a dataset need to be to profit from this treatment? And is there a dataset too big for `cortecs`?

Otherwise, I have a few fairly high-level comments on the paper and the docs. I have tried to stay away from commenting on small details for now.

### Paper:
The text of the paper is clear and well-written, however I would appreciate a more extensive description of the methods. Meanwhile, since the benchmark only shows downstream analysis (posterior samples), it would be nice to see both some additional examples, and more detail on the examples. For example, I would find it very interesting to have a figure or two showing how the reconstructed opacities compare to the original data (including the residuals) for different choices of compression. For example, for both rich and sparse opacities, how much does the choice of the number of components affect the quality of the PCA reconstruction, or the choice of hyperparameters for the neural network (or even architecture)? Could you also show an extreme example where `cortecs` actually fails? This would illustrate when `cortecs` is effective and what its limitations are so that users can get a better understanding of how to use it best.

I also noticed that there is no description of the compression algorithms in the paper. This is fine for PCA and the polynomials, but some description of the architectures of networks that are available and what the default choices are would be very helpful.

In figure one, I noticed that the posteriors without compression show much more substructure than the ones with compression. Is this because some behaviour has been removed by the compression, resulting in fewer regions of the prior volume that `multinest` finds with high likelihood? Or is it because the run with compression was faster per iteration, and hence you could generate more samples for a smoother posterior?

Docs:
I attempted to run the quickstart example, and found it did not load the data because the data appeared to have been saved with the keyword `allow_pickle=True`, however I do not seem to be able to set this keyword myself. As a result, I could not continue with the test. Since the same dataset is used for all the examples, I could not run any of them. As a result, I was also unable to verify the functionality of the package for the moment.

I would really like to see more detailed API documentation. What are the meanings of the parameters, how does changing them affect the output, etc? There are even some modules entirely missing, such as the optimisers.

Similar to the paper, the documentation does not describe what architecture(s) of neural networks can be used. What is the default network (fully-connected, I assume)? Can the user supply a different network architecture if they want to, and if so how must it be defined? How might the choice of network architecture affect the results?

I find the discussion of Optimisation of the fitting a bit difficult to follow, because some key details are missing. What is being optimised for - reconstruction accuracy, compression ratio, training/validation loss, … ? What is the algorithm being deployed for optimisation? Without this information, it is hard to tell when a users should try optimisation and what they should hope to get out of it.

### Code:
I think that @doriannblain has already covered everything required on the code itself, so I won’t make any comment on it right now.
