+++ src/cortecs/tests/test_opac.py
@@ -26,18 +26,14 @@ class TestOpac(unittest.TestCase):
         :return:
         """
         with self.assertRaises(ValueError):
-            opac_test = Opac_cia(
-                self.cia_filename, loader="wrongmethod", view="full_frame"
-            )
+            _ = Opac_cia(self.cia_filename, loader="wrongmethod", view="full_frame")
 
     def test_cia_opac_instantiated(self):
         """
         does an error get thrown when I'm just trying to make a little opacCIA object?
         :return:
         """
-        opac_test = Opac_cia(
-            self.cia_filename, loader="exotransmit_cia", view="full_frame"
-        )
+        _ = Opac_cia(self.cia_filename, loader="exotransmit_cia", view="full_frame")
         self.assertTrue(True)
 
     def test_cia_opac_join(self):
