### 2024-01-18 a19d975
optional neural network installation
