+++ paper/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 The absorption and emission of light by exoplanet atmospheres encode details of atmospheric
 composition, temperature, and dynamics. Fundamentally, simulating these processes requires detailed knowledge
 of the opacity of gases within an atmosphere. When modeling broad wavelength ranges at high resolution, such opacity data for
-even a single gas can take up multiple gigabytes of system memory. This aspect can be a limiting
+even a single gas can take up multiple gigabytes of system random-access memory (RAM). This aspect can be a limiting
 factor in determining the number of gases to consider in a simulation, or even in choosing the architecture of
 the system used for the simulation. Here, we present `cortecs`, a Python tool for compressing
 opacity data. `cortecs` provides flexible methods for fitting the
@@ -40,7 +40,7 @@ available for download with `pip`.
 
 # Statement of need
 Observations with the latest high-resolution spectrographs [e.g., @mace:2018; @seifahrt:2020; @pepe:2021]
-have motivated memory-intensive simulations of exoplanet atmospheres at high spectral resolution.
+have motivated RAM-intensive simulations of exoplanet atmospheres at high spectral resolution.
 `cortecs` enables these simulations with more
 gases and on a broader range of computing architectures by compressing opacity data.
 
@@ -49,14 +49,15 @@ radiative transfer equation over tens of thousands of wavelength points [e.g., @
 @beltz:2023; @line:2021; @wardenier:2023;  @gandhi:2023; @prinoth:2023; @maguire:2023].
 To decrease computational runtime,
 some codes have parallelized the problem on GPUs [e.g., @line:2021; @lee:2022]. However, GPUs cannot in general hold large amounts of
-data in-memory [e.g., @ito:2017]; only the cutting-edge, most expensive GPUs have memory in excess of 30 GB
-(such as the NVIDIA A100 or H100). Memory management is therefore a clear concern when producing
+data their video random-access memory (VRAM) [e.g., @ito:2017]; only the cutting-edge, most expensive GPUs are equipped with VRAM in excess of 30 GB
+(such as the NVIDIA A100 or H100). RAM and VRAM management is therefore a clear concern when producing
 high-resolution spectra.
 
-How do we decrease the memory footprint of these calculations? By far the largest contributor to the memory footprint,
+How do we decrease the RAM footprint of these calculations? By far the largest contributor to the RAM footprint,
 at least as measured on disk, is the opacity data. For instance, the opacity data for a single gas species across
-the wavelength range of the IGRINS spectrograph [@mace:2018] takes up 2.5 GB of memory at float64 precision and at a resolution of 400,000. It stands to reason
-that decreasing the amount of memory consumed by opacity data would strongly decrease the total amount of memory consumed
+the wavelength range of the IGRINS spectrograph [@mace:2018] takes up 2.5 GB of non-volatile memory (i.e., the file size is 2.5 GB) at float64 precision and at a resolving power of 400,000
+(as used in [@line:2021], with 39 temperature points and 18 pressure points, using, e.g., the [@polyansky:2018] water opacity tables). It stands to reason
+that decreasing the amount of RAM/VRAM consumed by opacity data would strongly decrease the total amount of RAM/VRAM consumed
 by the radiative transfer calculation.
 
 One solution is to isolate redundancy: While the wavelength dependence of opacity is sharp for many gases,
@@ -65,31 +66,85 @@ the temperature and pressure dependencies are generally smooth and similar acros
 This feature implies that the opacity data should be compressible without significant loss of
 accuracy at the spectrum level.
 
+While our benchmark case (see Benchmark) demonstrates the applicability of `cortecs` to high-resolution
+opacity functions of molecular gas, the package is general and can be applied to any opacity data that has
[…]
