+++ paper/paper.bib
@@ -241,6 +241,12 @@ archivePrefix = {arXiv},
   year={2023},
   publisher={Oxford University Press}
 }
+@misc{chollet:2015,
+  title={Keras},
+  author={Chollet, Fran\c{c}ois and others},
+  year={2015},
+  howpublished={\url{https://keras.io}},
+}
 
 @article{polyansky:2018,
   title={ExoMol molecular line lists XXX: a complete high-accuracy line list for water},
