### 2024-01-18 9a6898a
add installation check
