### 2024-01-18 7987877
add bounds checks for eval
