### 2024-01-18 977b583
add correct args to eval the nns
