+++ src/cortecs/eval/eval_neural_net.py
@@ -59,7 +59,9 @@ def feed_forward(x, n_layers, weights, biases):
 
 
 # @jax.jit
-def eval_neural_net(T, P, n_layers, weights, biases):
+def eval_neural_net(
+    T, P, temperatures, pressures, wavelengths, n_layers=None, weights=None, biases=None
+):
     """
     evaluates the neural network at a given temperature and pressure.
 
