### 2024-01-18 8906ff5
more descriptive docstring for eval_polynomial
