### 2024-01-18 5b192a2
more descriptive docstring for the neural nets
