### 2024-01-18 da0ffdd
more descriptive docstrings for PCA
