+++ src/cortecs/fit/fit_pca.py
@@ -1,4 +1,6 @@
-import warnings
+"""
+Module for performing PCA regression on opacity functions.
+"""
 
 import numpy as np
 
@@ -9,11 +11,11 @@ def standardize_cube(input_array):
 
     Inputs
     -------
-        :input_array: (n_exp x pixels) it's an array.
+        :input_array: (ntemperature x npressure) array for PCA.
 
     Returns
     -------
-        :standardized_cube: (n_exp x pixels) standardized array for PCA.
+        :standardized_cube: (ntemperature x npressure) standardized (mean-substracted, standard deviation-divided) array for PCA.
     """
     nf, nx = input_array.shape
     mat = input_array.copy()
@@ -30,7 +32,14 @@ def standardize_cube(input_array):
 
 def do_svd(standardized_cube, nc, nx):
     """
-    Does the SVD on the standardized flux cube.
+    Does the SVD (singular value decomposition) on the standardized cube.
+
+    Inputs
+    -------
+        :standardized_cube: (ntemperature x npressure) standardized (mean-substracted, standard deviation-divided) array for PCA.
+        :nc: (int) number of PCA components to keep in the reconstruction. Increasing the number of components can make the reconstruction
+        of opacity data more accurate, but it can also lead to overfitting and make the model size larger (i.e.,
+        decrease the compression factor).
 
     """
 
@@ -44,7 +53,19 @@ def do_svd(standardized_cube, nc, nx):
 
 def fit_mlr(cube, X):
     """
-    Fits the MLR to the flux cube with the PCA outputs and normalizes accordingly.
+    todo: check array shapes.
+    Fits the MLR (multiple linear regression) to the input cube with the PCA outputs and normalizes accordingly.
+
+    Inputs
+    -------
+        :cube: (ntemp x npressure) array being fit with PCA. Generally, this is the opacity data at a single
+        wavelength as a function of temperature and pressure.
+        :X: (npressure x ncomponents) PCA components. these are set by functions such as `do_pca`.
+
+    Returns
+    -------
+        :beta: (ncomponents x ntemp) PCA coefficients. These are the coefficients that are used to reconstruct the
+        opacity data at a later stage (e.g., within an atmospheric forward model). They represent the "compressed" opacity.
     """
     Y = np.moveaxis(cube.copy(), 0, -1)
     term1 = np.linalg.inv(np.dot(X.T, X))
@@ -59,9 +80,11 @@ def do_pca(cube, nc=3):
 
     Inputs
     -------
-        :wav_for_pca: (n_exp x pixels) it's an array.
-        :flux_for_pca: it's the same.
-        :nc: (int) number of PCA components to keep.
+        :cube: (ntemperature x npressure) array being fit with PCA. Generally, this is the opacity data at a single
+        wavelength as a function of temperature and pressure.
+        :nc: (int) number of PCA components to keep. Increasing the number of components can make the reconstruction
+        of opacity data more accurate, but it can also lead to overfitting and make the model size larger (i.e.,
+        decrease the compression factor).
 
     """
 
@@ -81,17 +104,20 @@ def do_pca(cube, nc=3):
     return xMat, standardized_cube, s, vh, u
 
 
-def fit_pca(cross_section, P, T, xMat, nc=3, wav_ind=1):
+def fit_pca(cross_section, P, T, xMat, nc=3, wav_ind=1, savename=None):
     """
     Fits the PCA to the opacity data.
 
     Inputs
     -------
-        :cross_section: (n_exp x pixels) it's an array.
+        :cross_section: (ntemp x npressure) the array of cross-sections being fit.
         :P: pressure grid
         :T: temperature grid
-        :nc: (int) number of PCA components to keep.
[…]
