### 2024-01-18 c393e22
remove savename from the polynomials
