### 2024-01-18 5ce52dd
remove naked try clause
