+++ src/cortecs/opac/interpolate_cia.py
@@ -4,10 +4,10 @@ resolution and subsequently chunk it up.
 Example instructions / workflow:
 >>> reference_file = '../opacFe/opacFe.dat'
 >>> CIA_file = 'opacCIA.dat'
->>> interpolate_CIA(CIA_file, reference_file)
+>>> interpolate_cia(CIA_file, reference_file)
 >>> CIA_file = 'opacCIA_highres.dat'
 >>> ref_file_base = '../opacFe/opacFe'
->>> chunk_wavelengths_CIA(CIA_file, ref_file_base)
+>>> chunk_wavelengths_cia(CIA_file, ref_file_base)
 And that should work!
 
 todo: generalize to PLATON CIA.
@@ -23,10 +23,10 @@ from tqdm import tqdm
 from cortecs.opac.chunking import *
 from cortecs.opac.io import *
 
-######################### Pt. 1: Interpolation ####################################
+# Pt. 1: Interpolation
 
 
-def check_temp_grid(df, real_temperature_grid, CIA_file):
+def check_temp_grid(df, real_temperature_grid, cia_file):
     """
     Checks that the temperature grid of the CIA file is the same as the reference file.
     Inputs
@@ -45,13 +45,14 @@ def check_temp_grid(df, real_temperature_grid, CIA_file):
     for temp in real_temperature_grid:
         if temp not in df.temp.unique():
             print(
-                f"Temperature {temp} not in CIA file {CIA_file}! Cannot interpolate in temperature yet. Will set these values to 0."
+                f"Temperature {temp} not in CIA file {cia_file}!"
+                + f" Cannot interpolate in temperature yet. Will set these values to 0."
             )
     return
 
 
-def interpolate_CIA(
-    CIA_file, reference_file, outfile=None, loader="exotransmit", load_kwargs={}
+def interpolate_cia(
+    cia_file, reference_file, outfile=None, loader="exotransmit", load_kwargs=None
 ):
     """
     Interpolates a CIA file to a higher resolution, using the wavelength grid
@@ -73,6 +74,8 @@ def interpolate_CIA(
         Creates a file with "hires" attached to the end of CIA_file that has been interpolated
         to higher resolution.
     """
+    if load_kwargs is None:
+        load_kwargs = {}
     reference_opac = Opac(reference_file, loader=loader, load_kwargs=load_kwargs)
 
     real_wavelength_grid = reference_opac.wl
@@ -82,14 +85,14 @@ def interpolate_CIA(
 
     df = Opac_cia(CIA_file, loader="exotransmit_cia", view="full_frame").cross_section
 
-    check_temp_grid(df, real_temperature_grid, CIA_file)
+    check_temp_grid(df, real_temperature_grid, cia_file)
 
     if reference_opac.wl.max() > df.wav.max() or reference_opac.wl.min() < df.wav.min():
         print(
             "Reference file has a larger wavelength grid than CIA file. Will fill with zeros."
         )
 
-    species_dict_interped = get_empty_species_dict(CIA_file)
+    species_dict_interped = get_empty_species_dict(cia_file)
     interped_wavelengths = []
     interped_temps = []
 
@@ -122,6 +125,7 @@ def interpolate_CIA(
     reference_species = species_dict_interped[list(species_dict_interped.keys())[0]]
 
     buffer = "   "  # there's a set of spaces between each string!
+    temp = np.nan  # fill value
     for i in tqdm(range(len(reference_species))):
         # the first line gets different treatment!
         if i == 0:
@@ -129,7 +133,7 @@ def interpolate_CIA(
                 interped_temps
             )  # add the LOWEST temperature in the temperature grid!
             new_string += ["{:.12e}".format(temp) + "\n"]
-        if interped_temps[i] != temp:
[…]
