### 2024-01-18 36ec035
actually implement optimize_pca
