### 2024-01-18 418e497
add nn installation instructions
