### 2024-01-18 b6e9f94
github tests with nn installation!
