+++ .github/workflows/python-package.yml
@@ -28,7 +28,7 @@ jobs:
         run: |
           python3 -m pip install -U pip
           python3 -m pip install -U setuptools setuptools_scm pep517
-          python3 -m pip install -e .
+          python3 -m pip install -e .[neural_networks]
 
 
       - name: Run tests
