### 2024-01-18 bac28f4
refactor name for interpolate_cia
