### 2024-01-18 bcb55c1
save nn fit correctly
