+++ src/cortecs/tests/test_integration.py
@@ -117,7 +117,7 @@ class TestIntegration(unittest.TestCase):
             / predictions.reshape(ntemp, npres)
         )
         median_err = np.median(np.abs(percent_errors))
-        fitter.save("test_nn")
+        cortecs.fit.fit_neural_net.save_neural_net("test_nn", [res])
         with open("test_nn.pkl", "rb") as f:
             all_weights, all_biases = pickle.load(f)
 
