### 2024-01-19 1ef4e2a
print those best params from the optimizer
