+++ src/cortecs/tests/test_integration.py
@@ -188,6 +188,7 @@ class TestIntegration(unittest.TestCase):
         max_size = 1.6
         max_evaluations = 8
         optimizer.optimize(max_size, max_evaluations)
+        print(optimizer.best_params)
         self.assertTrue(
             optimizer.best_params
             == {
