### 2024-01-19 f404d4c
update best activation func
