### 2024-01-19 5c7ce8c
pass correct args to eval nn
