### 2024-01-19 1f6e7d2
clean up complexity of writer_exotransmit_cia
