### 2024-01-19 00d9230
pass temp to writer and stuff
