### 2024-01-19 cd4c802
remove extraneous args from fit pca
