### 2024-01-19 5e15f98
pass kwargs to fit pca
