### 2024-01-19 b653451
remove an arg from fit nn
