### 2024-01-19 a00a4f5
remove an arg from calc_metrics
