### 2024-01-19 0f1db94
functionalize reused plotting code
