### 2024-01-19 14b8705
change that i do not ask for all_wl anymore
