### 2024-01-19 957a342
add temperature check to line break
