### 2024-01-19 209f9fb
reduce some cognitive complexity of interpolate_cia
