+++ src/cortecs/opac/interpolate_cia.py
@@ -51,6 +51,45 @@ def check_temp_grid(df, real_temperature_grid, cia_file):
     return
 
 
+def initialize_species_dict(species_dict, wavelength_grid):
+    """
+    Initializes a dictionary of species to be interpolated.
+    Inputs
+    ------
+        :species_dict: (dict) dictionary of species to be interpolated.
+    Outputs
+    -------
+        None
+    Side effects
+    -------------
+        Modifies species_dict in place.
+    """
+    for species_key in species_dict.keys():
+        species_dict[species_key] += list(np.ones_like(wavelength_grid) * 0.0)
+    return
+
+
+def add_line_string_species(line_string, species_dict, i, buffer):
+    """
+    Adds the species to the line string.
+    Inputs
+    ------
+        :line_string: (str) the line string to be modified.
+        :species_dict: (dict) dictionary of species to be interpolated.
+        :i: (int) index of the line string.
+        :buffer: (str) buffer between each species.
+    Outputs
+    -------
+        None
+    Side effects
+    -------------
+        Modifies line_string in place.
+    """
+    for species_key in species_dict_interped.keys():
+        line_string += "{:.12e}".format(species_dict_interped[species_key][i]) + buffer
+    return
+
+
 def interpolate_cia(
     cia_file, reference_file, outfile=None, loader="exotransmit", load_kwargs=None
 ):
@@ -99,10 +138,7 @@ def interpolate_cia(
     # perform interpolation
     for unique_temp in tqdm(real_temperature_grid, desc="Interpolating"):
         if unique_temp not in df.temp.unique():
-            for species_key in species_dict_interped.keys():
-                species_dict_interped[species_key] += list(
-                    np.ones_like(real_wavelength_grid) * 0.0
-                )
+            initialize_species_dict(species_dict_interped, real_wavelength_grid)
 
         else:
             sub_df = df[df.temp == unique_temp]
@@ -126,6 +162,7 @@ def interpolate_cia(
 
     buffer = "   "  # there's a set of spaces between each string!
     temp = np.nan  # fill value
+
     for i in tqdm(range(len(reference_species))):
         # the first line gets different treatment!
         if i == 0:
@@ -140,10 +177,7 @@ def interpolate_cia(
 
         line_string = wavelength_string + buffer
 
-        for species_key in species_dict_interped.keys():
-            line_string += (
-                "{:.12e}".format(species_dict_interped[species_key][i]) + buffer
-            )
+        add_line_string_species(line_string, species_dict_interped, i, buffer)
 
         new_string += [line_string + "\n"]
 
