### 2024-01-19 c46c6b1
actually define the dictionary I use!
