### 2024-01-19 539bf58
remove more complexity from interpolate_cia?
