### 2024-01-19 d817a7d
matrix the OS versions too! and more python 👹
