### 2024-01-19 f54d539
update versions of python setup and checkout
