### 2024-01-23 cdaf918
update plot with example application
