+++ paper/paper.bib
@@ -68,6 +68,16 @@ archivePrefix = {arXiv},
   year={2016},
   publisher={The Royal Society Publishing}
 }
+@article{@rothman:2010,
+  title={HITEMP, the high-temperature molecular spectroscopic database},
+  author={Rothman, Laurence S and Gordon, IE and Barber, RJ and Dothe, H and Gamache, Robert R and Goldman, A and Perevalov, VI and Tashkun, SA and Tennyson, J},
+  journal={Journal of Quantitative Spectroscopy and Radiative Transfer},
+  volume={111},
+  number={15},
+  pages={2139--2150},
+  year={2010},
+  publisher={Elsevier}
+}
 @article{alzubaidi:2021,
   title={Review of deep learning: Concepts, CNN architectures, challenges, applications, future directions},
   author={Alzubaidi, Laith and Zhang, Jinglan and Humaidi, Amjad J and Al-Dujaili, Ayad and Duan, Ye and Al-Shamma, Omran and Santamar{\'\i}a, Jos{\'e} and Fadhel, Mohammed A and Al-Amidie, Muthana and Farhan, Laith},
