+++ paper/paper.md
@@ -102,8 +102,11 @@ with a decompression method for evaluating opacity as a function of temperature,
 for GPUs and are accelerated with the `JAX` code transformation framework [@jax:2018]. An example of this reconstruction
 is shown in \autoref{fig:example}.
 
-![The original opacity functions (solid lines) and their `cortecs` reconstructions (dotted lines) over a large
-wavelength range and at multiple temperatures and pressures.\label{fig:example}](example_application.png)
+![Top panel: The original opacity function of CO [@rothman:2010] (solid lines) and its `cortecs` reconstruction (transparent lines) over a large
+wavelength range and at multiple temperatures and pressures. Bottom panel: the absolute residuals between the opacity function
+and its `cortecs` reconstruction. Note that opacities less than $10^-{60}$ are not generally relevant for the benchmark
+presented here; an opacity of $\sigma_\lambda=10^-{60}$ would require a column nearly $10^{27}$m long to become
+optically thick at a pressure of 1 bar and temperature of 1000 K. \label{fig:example}](example_application.png)
 
 
 # Workflow
