### 2024-01-23 7a58a58
remove python 3.12 testing. not sure why it always breaks!
