### 2024-01-23 9dd6928
metrics no longer needs opac passed
