### 2024-01-23 a153a55
update paper with table
