### 2024-01-23 c002395
clean up eval functions for nn
