+++ paper/paper.bib
@@ -9,6 +9,16 @@
   year={2021},
   publisher={Nature Publishing Group UK London}
 }
+@article{tennyson:2016,
+  title={The ExoMol database: molecular line lists for exoplanet and other hot atmospheres},
+  author={Tennyson, Jonathan and Yurchenko, Sergei N and Al-Refaie, Ahmed F and Barton, Emma J and Chubb, Katy L and Coles, Phillip A and Diamantopoulou, S and Gorman, Maire N and Hill, Christian and Lam, Aden Z and others},
+  journal={Journal of Molecular Spectroscopy},
+  volume={327},
+  pages={73--94},
+  year={2016},
+  doi={10.1016/j.jms.2016.05.002},
+  publisher={Elsevier}
+}
 @article{mansfield:2022,
   title={Confirmation of water absorption in the thermal emission spectrum of the hot Jupiter WASP-77Ab with HST/WFC3},
   author={Mansfield, Megan and Wiser, Lindsey and Stevenson, Kevin B and Smith, Peter and Line, Michael R and Bean, Jacob L and Fortney, Jonathan J and Parmentier, Vivien and Kempton, Eliza M-R and Arcangeli, Jacob and others},
@@ -44,7 +54,7 @@ archivePrefix = {arXiv},
   pages={A125},
   year={2014},
    doi = {10.1051/0004-6361/201322971},
-  publisher={EDP Sciences}
+  publisher={EDP Sciences},
 }
 @article{line:2013,
   title={A systematic retrieval analysis of secondary eclipse spectra. I. A comparison of atmospheric retrieval techniques},
@@ -55,7 +65,7 @@ archivePrefix = {arXiv},
   doi = {10.1088/0004-637X/775/2/137},
   pages={137},
   year={2013},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
 }
 @article{jolliffe:2016,
   title={Principal component analysis: a review and recent developments},
@@ -66,7 +76,17 @@ archivePrefix = {arXiv},
   doi = {10.1098/rsta.2015.0202},
   pages={20150202},
   year={2016},
-  publisher={The Royal Society Publishing}
+  publisher={The Royal Society Publishing},
+}
+@article{gordon:2017,
+  title={The HITRAN2016 molecular spectroscopic database},
+  author={Gordon, Iouli E and Rothman, Laurence S and Hill, Christian and Kochanov, Roman V and Tan, Y and Bernath, Peter F and Birk, Manfred and Boudon, V and Campargue, Alain and Chance, KV and others},
+  journal={Journal of Quantitative Spectroscopy and Radiative Transfer},
+  volume={203},
+  pages={3--69},
+  doi={10.1016/j.jqsrt.2017.06.038},
+  year={2017},
+  publisher={Elsevier}
 }
 @article{@rothman:2010,
   title={HITEMP, the high-temperature molecular spectroscopic database},
@@ -75,8 +95,9 @@ archivePrefix = {arXiv},
   volume={111},
   number={15},
   pages={2139--2150},
+  doi = {10.1016/j.jqsrt.2010.05.001},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
 }
 @article{alzubaidi:2021,
   title={Review of deep learning: Concepts, CNN architectures, challenges, applications, future directions},
@@ -86,7 +107,7 @@ archivePrefix = {arXiv},
   doi = {10.1186/s40537-021-00444-8},
   pages={1--74},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
 }
 @software{jax:2018,
[…]
