+++ paper/paper.bib
@@ -248,8 +248,60 @@ archivePrefix = {arXiv},
        eprint = {2308.04523},
  primaryClass = {astro-ph.EP},
 }
-
-
+@article{coles:2019,
+  title={ExoMol molecular line lists--XXXV. A rotation-vibration line list for hot ammonia},
+  author={Coles, Phillip A and Yurchenko, Sergei N and Tennyson, Jonathan},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={490},
+  number={4},
+  pages={4638--4647},
+  year={2019},
+  doi={10.1093/mnras/stz2778},
+  publisher={Oxford University Press}
+}
+@article{karman:2019,
+  title={Update of the HITRAN collision-induced absorption section},
+  author={Karman, Tijs and Gordon, Iouli E and van Der Avoird, Ad and Baranov, Yury I and Boulet, Christian and Drouin, Brian J and Groenenboom, Gerrit C and Gustafsson, Magnus and Hartmann, Jean-Michel and Kurucz, Robert L and others},
+  journal={Icarus},
+  volume={328},
+  pages={160--175},
+  year={2019},
+  doi={10.1016/j.icarus.2019.02.034},
+  publisher={Elsevier}
+}
+@article{barber2014exomol,
+  title={ExoMol line lists--III. An improved hot rotation-vibration line list for HCN and HNC},
+  author={Barber, RJ and Strange, JK and Hill, C and Polyansky, OL and Mellau, G Ch and Yurchenko, SN and Tennyson, Jonathan},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={437},
+  number={2},
+  pages={1828--1835},
+  year={2014},
+  doi={10.1093/mnras/stt2011},
+  publisher={Oxford University Press}
+}
+@article{azzam:2016,
+  title={ExoMol molecular line lists--XVI. The rotation--vibration spectrum of hot H2S},
+  author={Azzam, Ala'a AA and Tennyson, Jonathan and Yurchenko, Sergei N and Naumenko, Olga V},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={460},
+  number={4},
+  pages={4063--4074},
+  year={2016},
+  doi={10.1093/mnras/stw1133},
+  publisher={Oxford University Press}
+}
+@article{hargreaves:2020,
+  title={An accurate, extensive, and practical line list of methane for the HITEMP database},
+  author={Hargreaves, Robert J and Gordon, Iouli E and Rey, Michael and Nikitin, Andrei V and Tyuterev, Vladimir G and Kochanov, Roman V and Rothman, Laurence S},
+  journal={The Astrophysical Journal Supplement Series},
+  volume={247},
+  number={2},
+  pages={55},
+  year={2020},
+  doi={10.3847/1538-4365/ab7a1a},
+  publisher={IOP Publishing}
+}
 @article{gandhi:2023,
   title={Retrieval Survey of Metals in Six Ultrahot Jupiters: Trends in Chemistry, Rain-out, Ionization, and Atmospheric Dynamics},
   author={Gandhi, Siddharth and Kesseli, Aurora and Zhang, Yapeng and Louca, Amy and Snellen, Ignas and Brogi, Matteo and Miguel, Yamila and Casasayas-Barris, N{\'u}ria and Pelletier, Stefan and Landman, Rico and others},
