### 2024-01-24 f839e6a
make equations inline?
