+++ paper/paper.bib
@@ -88,7 +88,7 @@ archivePrefix = {arXiv},
   year={2017},
   publisher={Elsevier}
 }
-@article{@rothman:2010,
+@article{rothman:2010,
   title={HITEMP, the high-temperature molecular spectroscopic database},
   author={Rothman, Laurence S and Gordon, IE and Barber, RJ and Dothe, H and Gamache, Robert R and Goldman, A and Perevalov, VI and Tashkun, SA and Tennyson, J},
   journal={Journal of Quantitative Spectroscopy and Radiative Transfer},
