+++ paper/paper.bib
@@ -269,7 +269,7 @@ archivePrefix = {arXiv},
   doi={10.1016/j.icarus.2019.02.034},
   publisher={Elsevier}
 }
-@article{barber2014exomol,
+@article{barber:2014,
   title={ExoMol line lists--III. An improved hot rotation-vibration line list for HCN and HNC},
   author={Barber, RJ and Strange, JK and Hill, C and Polyansky, OL and Mellau, G Ch and Yurchenko, SN and Tennyson, Jonathan},
   journal={Monthly Notices of the Royal Astronomical Society},
