### 2024-01-24 #6104 @arjunsavel
Thank you to both referees for comprehensive and helpful responses! We’ve implemented the requested changes — see the below descriptions. Please let me know if you have any follow-up questions or concerns about the codebase. Thank you again!

**Response to @doriannblain**

> I have two main issues. The first one is that I find that the core functionality of cortecs (lowering the memory usage of opacities) is not well exposed in the documentation and lacks a more concrete example. My understanding is that a real-world usage of cortecs would be the following:

> Load opacities with an Opac object and fit them with a Fitter object,

> Save the "fitting parameters" on disk,

> When initializing a retrieval, load the previously saved fitting parameters,

> Use an eval function with the loaded fitting parameters to retrieve the opacities on-demand.

> Step 1 is already well described. I suggest that the authors rework the "Quickstart" section to include steps 2 to 4 in a clearer way. Step 3 can simply be the loading step. The "Checking accuracy and speed" subsection is useful. It should be kept, but at the end of the section (and added in some form to the paper, see below).

We have now included these steps in the tutorial and explicitly outlined these steps in the new “Workflow” section of the paper.

> My second main issue is that it is unclear what exactly are the limitations of cortecs (if there are any at all). The benchmark provided in the article, while convincing, only shows the case of a hot Jupiter. I acknowledge that this is a powerful example, as currently, most high-resolution studies are done on this kind of planet. At the same time, hot Jupiters emission spectra in the near-infrared (the benchmark case) usually have most of their contribution coming from a limited region of the opacity data pressure-temperature space, that is 800-1500 K and 0.01-10 bar, while opacity data pressure and temperature grid commonly extend from 100 to 3000 K, and 1e3 to 1e-6 bar. While I don't think more benchmarks are necessary, the article should quantify the differences between the "decompressed" opacities and the original opacities in the entirety of the wavelength-pressure-temperature space of the opacity data, at least for key species such as H2O and CO (this is already partially done in the documentation). A tool/function should also be provided to the users so they can easily evaluate these differences. In addition, it should be made clear in the documentation that this evaluation is part of the intended, standard cortecs workflow. This should also be assorted by a recommendation to compare spectra built with and without using cortecs.

Great point. Cortecs now includes a tool — cortecs.fit.metrics.calc_metrics — that easily calculates the absolute and percentage errors. We also have a stronger warning at the end of the quickstart.

> I obtained infinite values in the "Quickstart" example when I ran the eval function at some pressures and temperatures (e.g. at T[1], P[5] and wl[-375]). On my test on the last 500 wavelength points, there was more than 17000 infinite values (~1% of the tested range). This should be fixed.

We have addressed this bug — all values should be finite now.

> Still in the "Quickstart" PCA example, and related to my second main issue, the highest finite error I obtained was 70% when evaluating the whole opacity file (at T[0], P[12] and wl[3949]). This is not necessarily worrying as the authors show a convincing corner plot in their article, but I would like them to comment on this. Also, this should be highlighted in the example and in the article, instead of simply showing the 9% error at one point.

We now comment on the global fit quality, and we now include a warning that users check that the accuracy level is sufficient.

> The role of the nc argument in the PCA example should be explained, as it drives the "compression factor" of this method.

We have indicated that nc refers to the number of principal components. In a similar vein, we have also clarified the purpose of the wav_ind parameter.

> The units of the pressure, temperature and wavelength arrays should be indicated in the example.

Done!

> The installation test outlined in the documentation’s "Installation" section does not run any actual test. Instead, this could be as simple as verifying that python3 -c ‘import cortecs’ does not raise any error.

We have added the suggested text to the installation instructions.

> The authors make a good job at explaining the divide by zero warning occurring in the "Quickstart" section (cell 3). However, a better way is to prevent this warning from being raised altogether. I suggest replacing src/opac/io.py:157-160 with something like:

Thanks for this suggestion — we’ve now implemented it.

> In the same place, could the authors explain this choice of 1e-104 as a threshold value, instead of e.g. sys.float_info.min?

Setting these values too low could break the fits. Perhaps a better solution would be to mask out these  points in the fitting process. -104 is set as a boundary for standard opacity functions (e.g., as used by the PLATON code).

> Likewise, the error raised in cell 8 of the "Quickstart" section is well explained, but it would be better if that error is not raised at all (it also prevents an easy full rerun of the notebook). The code already has the correct condition to handle it, so why not return hyperparameters that would reconstruct constant opacities instead of raising an error? It would be much more practical for the user and expand the use cases of cortecs.

To allow an easy rerun of the whole notebook, we’ve ensured that the first one passed to the function actually works. We’ve additionally passed a “force_fit_constant” parameter to “prep_pca”, which allows fitting constant opacities. The code can currently fit constant opacity functions at certain wavelengths, temperatures, or pressures — it’s just that for fitting a opacity function O(t,P, lambda), our scheme is to find the components that fit O(t,P, 0), then perform linear regression to fit those components to O(t, P, lambda).

> what is the purpose of calling fit_mlr in do_pca when prep_pca only returns xMat, which is calculated by do_svd?

We thank the reviewer for a detailed review of our code. This is extraneous code that has now been omitted.

> what is the purpose of calculating fNorm in fit_mlr when fit_pca only returns beta?

Similarly, this is extraneous code that has now been omitted.

> Is there a reason for the SVD to be performed along the pressure grid (size 13) rather than the temperature grid (size 30)? Would not the "compression factor" be higher, up to 15 (or 30?) for nc=1, if the temperature grid was chosen instead?

Interesting point. We chose these axes because of how the arrays are shaped when fitting along multiple axes. The pca_coeffs array is shaped (N_WAVE, N_PC, N_PRES), and the vectors array is shaped (N_TEMP, N_PC). If we swapped which axis we fit along, then the PCA_COEFFS array would be larger. Naively, it seems like this would decrease the compression factor, though this may just be how we’re packaging our data?

> The Evaluator.eval function uses the temperature and pressure grid, so its arguments should be indices, not the actual temperature and pressure values.

Two of the eval funcs (neural network and polynomial) use the actual temperature–pressure values, so it’s quicker to pass those values in than index the array. It also seems a bit more user-intuitive to evaluate the opacity function at a temperature and pressure value. Happy to change if you disagree!

> Related to the above point, Evaluator.eval should raise a clear message when the given pressure or temperature are out-of-bounds.

We now raise temperature, pressure, and wavelength out-of-bounds errors.

> A useful addition is the ability for the user to easily save and load "fitting parameters" (as outlined in my general comment). Right now, this is implemented for neural networks but not for PCA fits, and it is inoperant for Polynomial fits (the save argument of fit_polynomial does not do anything). I suggest implementing this feature for all the fitting frameworks.

We have now implemented parameter-saving for the remaining fitting frameworks.

> Since Tensorflow is optional for this package to work, I suggest making it an optional dependency.

True, tensorflow is optional — and it can be a bit tricky to manage in an environment. We’ve made tensorflow and keras optional dependencies, as not all users will be performing neural network fits.

> The code has minor logic weaknesses, deprecated docstrings, and could better follow the PEP8 convention. A PyCharm code inspection gave me several warnings and more than 100 weak warnings, most of them being PEP8 violations. The authors can also use flake8 to catch those. These should be quickly fixable and improve the code quality.

We addressed the minor logic weaknesses. The remaining flake8 errors are because lines are too long or from star imports. We prefer to keep line lengths as they are (we’re conforming to the black code formatting style), but we can change to autopep8 if the reviewer desires.

> Some of the modules (e.g. fit_pca.py) have the if __name__ == '__main__' pattern, signalling them as scripts, which they should not be.

We have fixed this.

> Looking at the cortecs repository commits, it seems that A. B. Savel is the only contributor to the code. Can the authors expand on the contribution of M. Bedell and E. M.-R. Kempton?

Thank you for the careful read of the paper. M. Bedell and E. M. R. Kempton both advised heavily on the design of this project.

> I would suggest using more specific terms when referring to "memory" in the paper (i.e., make a clear distinction between RAM, VRAM, non-volatile memory), to prevent potential confusions.

Thank you for the suggestion — the paper reads more clearly now! We have replaced “memory” with specific qualifiers in all instances.

> l.17: the authors state that the code is available on conda, but the documentation states that it will be distributed to conda "at a later date". I do not think the package is on conda yet, can the authors solve this?

We have amended the line in the paper to reflect that this package is not currently distributed on conda.

> l.33: the authors refer to the memory footprint "on disk". It would also be interesting to at least briefly mention how the RAM/VRAM usage usually compares with disk usage, as this is the main issue that cortecs is addressing.

We address this as follows: “These on-disk memory quotes are relatively faithful to the in-memory RAM footprint of the data when stored as `numpy`arrays (2.1 GB for the uncompressed data and 160 MB for the compressed data).

> l.34: the IGRINS acronym should be defined.

We have clarified that IGRINS is a spectrograph and spelled out its acronym (Immersion GRating INfrared Spectrometer).

> l.34—35: can the authors give a reference for the opacity data used in this example, as well as specify the size of the pressure and temperature girds?

We have clarified the opacity tables used.

> l.35: I suggest using the more precise term "resolving power" instead of "resolution".

We have made the requested change.

> l.38—39: I suggest adding a figure illustrating the wavelength, pressure, and temperature dependencies of the opacities. This figure could also be used to show an example of a fit(s) with cortecs.

We’ve included a new figure showing wavelength, pressure, and temperature dependencies of the opacities along with residuals against cortecs fits.

> Could the authors discuss their choice of using the term "compression/decompression" and not for example "fitting/modelling"? If "compression" is appropriate, then it should be specified if the compression is lossy (which I believe to be the case) or lossless, and why.

We include the following text now: “We generally refer to this process as ``compression'' as opposed to ``fitting'' to emphasize that we do not seek to construct physically motivated, predictive, or comprehensible models of the opacity function. Rather, we simply seek representations of the opacity function that consume less RAM/VRAM. The compression methods we use are ``lossy,’’ in that the original opacity data cannot be exactly recovered with our methods. We find that the loss of accuracy is tolerable for at least the hot Jupiter application (see Benchmark below).”

> I think this section would benefit from a slightly more concrete/detailed explanation on how the "compression/decompression" is achieved, that is, saving and loading "fitting parameters" (I will use this term despite it not being exact for PCA and neural nets). A small example would also be valuable.

We have included a new “workflow” section that hopefully illustrates this more clearly. Happy to elaborate!

> l.52—55: the features detailed in this paragraph are not mentioned in the documentation. Besides, they are not fully implemented yet. However, these features seem peripheral. If that is indeed the case, implementing them is not necessary for this publication, and this paragraph could be removed. If those features are important, then examples for each of them should be provided in the documentation.

Referring to “In addition to these compression/decompression methods, `cortecs` provides utility scripts for working with large opacity files. For instance, `cortecs` can convert opacity files between popular formats, break opacity files into wavelength "chunks" for parallel computations across CPUs, and add overlap between chunked files for calculations that include Doppler shifts.”— we have removed this paragraph.

> the opacity data sizes before and after "compression" should be added, as well as a reference for the opacities used. In in that case, it would also be interesting to see the size of the PCA vectors and PCA coefficients. A "compression" performance comparison of each method would also be valuable (this can be e.g. a small table listing compression factors).

We now list the data sizes before and after compression. We also provide a table demonstrating the compression factor achieved using other methods. We have also provided a table demonstrating these sizes for the benchmark dataset. Would it be helpful if we provided a second table for the Quickstart dataset?

> in the "Quickstart" example, the combined size (i.e. number of elements) of those two arrays (PCA vectors and PCA coefficients) is 3.2 times smaller than the size of the cross-section array, so less than the factor 13 mentioned in the article (but still significant). Even by setting nc=1, this factor only reaches 6.5. I am probably missing something as 6.5 is suspiciously the half of 13, which is also the size of the pressure grid. In all cases, the setup leading to that 13 "compression factor" needs to be exposed. It should also be clarified that the maximum "compression factor" (for that method at least) depends on the size of the pressure grid, if that is indeed the case.

We have clarified the shape / size of the opacity data used in the benchmark. It is different from the Quickstart opacity data — i.e., it has 18 pressure points, as opposed to 13.

> l.66: the runtimes with "compressed" and "uncompressed" opacities should be indicated, accompanied by the hardware used. The data reading time, and data "decompression" time should also be indicated. Similarly to my previous comment, a performance comparison of all "compression" methods would be valuable.

We have included these timing estimates in the paper.

> Are the retrieval and data described here detailed in another (upcoming) article? If yes, this should be mentioned.

We now indicate this in the article: “The non-compressed retrieval uses the data and retrieval framework from [@line:2021]; these specific samples were generated for an upcoming article (Savel et al. 2024, submitted)”

> While convincing, Fig. 1 by itself is not sufficient to claim that cortecs can be used "in [all] high-resolution retrievals". For that, the article needs to display examples challenging the opacity evaluating functions, such as very cold or very hot planets, planets with a thin atmosphere, or planets with high (Z > 100) or extreme metallicity (e.g. steam, Earth-like or Venus-like atmospheres). I think this is beyond the scope of this paper though, so I would suggest instead precising "in at least some high-resolution retrievals". Adding for example a histogram showing the differences between the "decompressed" and "uncompressed" opacities (akin to what is done in the documentation), and a small discussion around it, could help the reader forming an informed opinion.

We have included the suggested wording, and we add the fitting results per a suggestion above.

> The gold and teal labels are misaligned; this should be fixed.

We have fixed the alignment of the labels.

> In the caption: WASP-77A -> WASP-77Ab.

Done!

> The documentation mentions an Optimizer object that looks partially implemented in the code (it seems to only work for neural nets) and not mentioned in the paper. Do the authors have plans to develop on this feature in this submission?

Yes, we plan to develop on this feature. We have now extended cortecs to cover the PCA method, as well. Doing so for the polynomial method would require a significant testing refactor — we’re happy to do so if it would make the case for Optimizer stronger.

**Response to @pscicluna**

> Similar to [@doriannblain](https://github.com/doriannblain), I also find it a little difficult to see how cortecs would be integrated into an analysis pipeline. I imagine the idea is to start your pipeline by generating the compressed opacities, then (if necessary) distributing the evaluators to all the compute devices, then the sampler can run on all devices to get the posterior you’re after? If that’s the case, it would be a good idea to provide an example workflow, either in the docs or by describing it further in the paper.

We now describe an example workflow after the methods section.

> I also find it hard to tell when cortecs can be deployed to maximum benefit, and when it should be avoided. For example, can it only be used for molecular gas opacities, or could it be used to compress atomic opacities, or even temperature-dependent dust opacities? How large and complex does a dataset need to be to profit from this treatment? And is there a dataset too big for cortecs?

We describe in detail a number of potential failure modes at the end of the “Statement of Need” section. We are happy to mock up a toy model *showing* such failure if you’d like!

> The text of the paper is clear and well-written, however I would appreciate a more extensive description of the methods. Meanwhile, since the benchmark only shows downstream analysis (posterior samples), it would be nice to see both some additional examples, and more detail on the examples. For example, I would find it very interesting to have a figure or two showing how the reconstructed opacities compare to the original data (including the residuals) for different choices of compression. For example, for both rich and sparse opacities, how much does the choice of the number of components affect the quality of the PCA reconstruction, or the choice of hyperparameters for the neural network (or even architecture)? Could you also show an extreme example where cortecs actually fails? This would illustrate when cortecs is effective and what its limitations are so that users can get a better understanding of how to use it best.

We include a new plot (new Fig. 1) showing temperature, pressure, and wavelength dependencies. The failure point of cortecs are described in response to the second comment of this review

[…]
