### 2024-02-01 #6104 @pscicluna
Thanks for your speedy updates! The new additions markedly improve the paper. I do have some more comments, which hopefully will be fairly minor. I am, however, still having issues with the tutorials (see below) and will likely have more comments after that is fixed, as I have not really been able to interact with `cortecs` as yet.

The description of potential failure modes is very helpful. The authors' suggestion of a figure showing it would be greatly appreciated! Seeing how it fails will help users diagnose their own issues. Adding this to the documentation as well would be great.

The idea behind Figure 1 is definitely what I had in mind. However, I find it a bit difficult to read - the opacity spectra are so rich that it is hard to see any details. It is also unclear from the figure which compression method was used. It might be helpful to enlarge the figure and/or zoom in on narrow wavelength ranges of the spectra. I realise that the residual plot shows that the differences are generally small except in extreme cases, but seeing the details would be much more illustrative of the power of `cortecs` than the current plot it. Ideally, the results of multiple different compression methods would be shown on the same data, to give a sense of the trade-offs between them.

The improved API docs are greatly appreciated - however, it appears that these changes have not yet synced to RTD. I assume this is because no new version has been tagged yet, in which case it's not a big deal.

I still find the description of some details a bit lacking. For example, the optimisation is incomplete. I see from the code that it is performing a grid-search to find the hyperparameters with the lowest MSE, but it would be helpful to have this described in the documentation. It's also not clear to me from the code how the maximum size of the compressed data is considered in the optimisation - the code to me looks like it is just performing a grid search over the hyperparameters, and not considering the size of the compressed data.

I noticed that the package requires that devs update the version number manually. This can be easy to forget, and result in a mismatch between tags and the version number in the code. It may be worth considering using `setuptools-git-versioning` and making the version a dynamic value in `pyproject.toml`. This ensures that the git tag (or even commit hash) is the ultimate arbiter of version numbers, and can help prevent mismatches. Since this would also involve updating the package from `setup.py` to `pyproject.toml`, it may be worth either waiting until the next major release, or trying to use `python-versioneer` instead of `setuptools-git-versioning`.

I'm still facing issues with loading the data for the quickstart tutorial. This may be related to my environment, but I can't be sure.
Running on Colab with numpy==1.23.5 I receive the following error:

```python
UnpicklingError                           Traceback (most recent call last)
/usr/local/lib/python3.10/dist-packages/numpy/lib/npyio.py in load(file, mmap_mode, allow_pickle, fix_imports, encoding, max_header_size)
    440             try:
--> 441                 return pickle.load(fid, **pickle_kwargs)
    442             except Exception as e:

UnpicklingError: invalid load key, '{'.

UnpicklingError                           Traceback (most recent call last)
3 frames
<ipython-input-5-98289f3fc416> in <cell line: 1>()
----> 1 opac_obj = Opac(cross_sec_filename, loader="platon", load_kwargs=load_kwargs)
      2
      3 opac_obj

/usr/local/lib/python3.10/dist-packages/cortecs/opac/opac.py in __init__(self, filename, loader, load_kwargs)
     46         self.filename = filename
     47         self.load_obj = self._get_loader(loader)
---> 48         self.wl, self.T, self.P, self.cross_section = self.load_obj.load(
     49             filename, **load_kwargs
     50         )

/usr/local/lib/python3.10/dist-packages/cortecs/opac/io.py in load(self, cross_sec_filename, T_filename, P_filename, wl_filename)
    139         # temperatures are in K.
    140         # pressures are in Pa, I believe.
--> 141         wl = np.load(wl_filename, allow_pickle=True)
    142         T = np.load(T_filename, allow_pickle=True)
    143         P = np.load(P_filename, allow_pickle=True)

/usr/local/lib/python3.10/dist-packages/numpy/lib/npyio.py in load(file, mmap_mode, allow_pickle, fix_imports, encoding, max_header_size)
    441                 return pickle.load(fid, **pickle_kwargs)
    442             except Exception as e:
--> 443                 raise pickle.UnpicklingError(
    444                     f"Failed to interpret file {file!r} as a pickle") from e
    445

UnpicklingError: Failed to interpret file 'wavelengths.npy' as a pickle
```

However, running on Windows with python==3.11.7 and numpy==1.26.3 I get a different one:

```python
UnpicklingError                           Traceback (most recent call last)
File ...anaconda3\envs\cortecs_review\Lib\site-packages\numpy\lib\npyio.py:465, in load(file, mmap_mode, allow_pickle, fix_imports, encoding, max_header_size)
    464 try:
--> 465     return pickle.load(fid, **pickle_kwargs)
    466 except Exception as e:

UnpicklingError: invalid load key, '\x0a'.

UnpicklingError                           Traceback (most recent call last)
Cell In[7], line 1
----> 1 opac_obj = Opac(cross_sec_filename, loader="platon", load_kwargs=load_kwargs)
      3 opac_obj

File ...anaconda3\envs\cortecs_review\Lib\site-packages\cortecs\opac\opac.py:48, in Opac.__init__(self, filename, loader, load_kwargs)
     46 self.filename = filename
     47 self.load_obj = self._get_loader(loader)
---> 48 self.wl, self.T, self.P, self.cross_section = self.load_obj.load(
     49     filename, **load_kwargs
     50 )
     52 self.n_wav, self.n_t, self.n_p = len(self.wl), len(self.T), len(self.P)
     54 return

File ...anaconda3\envs\cortecs_review\Lib\site-packages\cortecs\opac\io.py:141, in loader_platon.load(self, cross_sec_filename, T_filename, P_filename, wl_filename)
    120 """
    121 loads in opacity data that's built for PLATON. To be passed on to Opac object.
    122
   (...)
    136     name of wavelength file
    137 """
    138 # todo: check wl units. They're in meters here.
    139 # temperatures are in K.
    140 # pressures are in Pa, I believe.
--> 141 wl = np.load(wl_filename, allow_pickle=True)
    142 T = np.load(T_filename, allow_pickle=True)
    143 P = np.load(P_filename, allow_pickle=True)

File ...anaconda3\envs\cortecs_review\Lib\site-packages\numpy\lib\npyio.py:467, in load(file, mmap_mode, allow_pickle, fix_imports, encoding, max_header_size)
    465     return pickle.load(fid, **pickle_kwargs)
    466 except Exception as e:
--> 467     raise pickle.UnpicklingError(
    468         f"Failed to interpret file {file!r} as a pickle") from e

UnpicklingError: Failed to interpret file 'wavelengths.npy' as a pickle
```

As you can see, there are strong similarities between the two errors, but the details are slightly different. However, I wonder if the larger problem is that pickle is not really intended to be completely portable? It might be better to use a more portable format, perhaps HDF5, to package the example data. For all its faults, even FITS would guarantee that the data would be widely portable.

### Minor comments:
There are some small typesetting issues in the caption of Figure 1. It looks like some of the math-mode delimiters are missing, and as a result the exponents are not being rendered correctly.

Line 112/113: "These opacity data *(cite them)* were originally stored" my emphasis - please insert the citation here. Or, if is the same data mentioned in the previous paragraph, make that clear.
