### 2024-02-01 565ad3e
update math mode in paper
