### 2024-02-01 5dfbe0b
wrap around no install of keras
