### 2024-02-01 37358f1
explain floor of opacity
