### 2024-02-05 0c201f5
add matplotlib to doc requirements so that RTD can use autodoc
