+++ docs/doc_requirements.txt
@@ -10,3 +10,4 @@ keras
 tensorflow
 h5py
 mock
+matplotlib
