+++ paper/paper.bib
@@ -1,3 +1,36 @@
+@article{kurucz2017including,
+  title={Including all the lines: data releases for spectra and opacities},
+  author={Kurucz, Robert L},
+  journal={Canadian Journal of Physics},
+  volume={95},
+  number={9},
+  pages={825--827},
+  year={2017},
+  publisher={NRC Research Press},
+  doi={10.1139/cjp-2016-0794},
+}
+@article{showman2020atmospheric,
+  title={Atmospheric dynamics of hot giant planets and brown dwarfs},
+  author={Showman, Adam P and Tan, Xianyu and Parmentier, Vivien},
+  journal={Space Science Reviews},
+  volume={216},
+  pages={1--83},
+  year={2020},
+  publisher={Springer},
+  doi={10.1007/s11214-020-00758-8}
+}
+@article{barstow2020comparison,
+  title={A comparison of exoplanet spectroscopic retrieval tools},
+  author={Barstow, Joanna K and Changeat, Quentin and Garland, Ryan and Line, Michael R and Rocchetto, Marco and Waldmann, Ingo P},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={493},
+  number={4},
+  pages={4884--4909},
+  year={2020},
+  publisher={Oxford University Press},
+  doi={doi.org/10.1093/mnras/staa548}
+}
+
 @article{line:2021,
   title={A solar C/O and sub-solar metallicity in a hot Jupiter atmosphere},
   author={Line, Michael R and Brogi, Matteo and Bean, Jacob L and Gandhi, Siddharth and Zalesky, Joseph and Parmentier, Vivien and Smith, Peter and Mace, Gregory N and Mansfield, Megan and Kempton, Eliza M-R and others},
