### 2024-02-05 69efc2d
update paper plots and metric API
