### 2024-02-05 b90dbd6
fix quotes and latex
