+++ .github/workflows/python-package.yml
@@ -39,7 +39,7 @@ jobs:
          pytest --cov=src/cortecs --cov-report=xml
 
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v4
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
