+++ .github/workflows/python-package.yml
@@ -43,11 +43,8 @@ jobs:
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
-          directory: ./coverage/reports/
-          env_vars: OS,PYTHON
           fail_ci_if_error: true
-          files: ./coverage.xml
+          files: ./coverage1.xml,./coverage2.xml # optional
           flags: unittests
           name: codecov-umbrella
-          path_to_write_report: ./coverage/codecov_report.txt
           verbose: true
