### 2024-02-07 711b733
add failure mode
