### 2024-02-07 3c8e983
add .txt files in case .npy does not work
