+++ docs/pages/wavelengths.txt
@@ -1,0 +1,4616 @@
+3.000000000000000923e-07
+3.002995104202230821e-07
+3.005993198620855132e-07
+3.008994286241209714e-07
+3.011998370051605629e-07
+3.015005453043356142e-07
+3.018015538210741780e-07
+3.021028628551038389e-07
+3.024044727064515019e-07
+3.027063836754436041e-07
+3.030085960627056909e-07
+3.033111101691654339e-07
+3.036139262960489779e-07
+3.039170447448839056e-07
+3.042204658174988139e-07
+3.045241898160236319e-07
+3.048282170428898322e-07
+3.051325478008303254e-07
+3.054371823928821597e-07
+3.057421211223828684e-07
+3.060473642929736461e-07
+3.063529122085987663e-07
+3.066587651735058462e-07
+3.069649234922458467e-07
+3.072713874696754547e-07
+3.075781574109539065e-07
+3.078852336215458469e-07
+3.081926164072206407e-07
+3.085003060740531140e-07
+3.088083029284236601e-07
+3.091166072770178687e-07
+3.094252194268293320e-07
+3.097341396851564152e-07
+3.100433683596047446e-07
+3.103529057580872079e-07
+3.106627521888240068e-07
+3.109729079603423924e-07
+3.112833733814796300e-07
+3.115941487613793461e-07
+3.119052344094944929e-07
+3.122166306355869780e-07
+3.125283377497278761e-07
+3.128403560622979053e-07
+3.131526858839870564e-07
+3.134653275257972933e-07
+3.137782812990392704e-07
+3.140915475153349797e-07
+3.144051264866176976e-07
+3.147190185251319323e-07
+3.150332239434333709e-07
+3.153477430543917376e-07
+3.156625761711870357e-07
+3.159777236073127234e-07
+3.162931856765751318e-07
+3.166089626930939942e-07
+3.169250549713026048e-07
+3.172414628259474484e-07
+3.175581865720913236e-07
+3.178752265251095310e-07
+3.181925830006928914e-07
+3.185102563148474803e-07
+3.188282467838947873e-07
+3.191465547244715041e-07
+3.194651804535323303e-07
+3.197841242883464793e-07
+3.201033865465006432e-07
+3.204229675458984634e-07
+3.207428676047610064e-07
+3.210630870416271354e-07
+3.213836261753528742e-07
+3.217044853251147957e-07
+3.220256648104059455e-07
+3.223471649510391241e-07
+3.226689860671463576e-07
+3.229911284791793741e-07
+3.233135925079097094e-07
+3.236363784744285487e-07
+3.239594867001496376e-07
+3.242829175068055768e-07
+3.246066712164508926e-07
+3.249307481514616658e-07
+3.252551486345357444e-07
+3.255798729886924777e-07
+3.259049215372757880e-07
+3.262302946039503581e-07
+3.265559925127047020e-07
+3.268820155878507418e-07
+3.272083641540242306e-07
+3.275350385361850706e-07
+3.278620390596168362e-07
+3.281893660499298980e-07
+3.285170198330576636e-07
+3.288450007352597013e-07
+3.291733090831211577e-07
+3.295019452035534460e-07
+3.298309094237934514e-07
+3.301602020714069731e-07
+3.304898234742847001e-07
+3.308197739606453879e-07
+3.311500538590354878e-07
+3.314806634983294119e-07
+3.318116032077300089e-07
+3.321428733167680356e-07
+3.324744741553053325e-07
+3.328064060535309068e-07
+3.331386693419642141e-07
+3.334712643514546826e-07
+3.338041914131820303e-07
+3.341374508586558414e-07
+3.344710430197189021e-07
+3.348049682285431234e-07
[…]
