+++ docs/pages/temperatures.txt
@@ -1,0 +1,30 @@
+1.000000000000000000e+02
+2.000000000000000000e+02
+3.000000000000000000e+02
+4.000000000000000000e+02
+5.000000000000000000e+02
+6.000000000000000000e+02
+7.000000000000000000e+02
+8.000000000000000000e+02
+9.000000000000000000e+02
+1.000000000000000000e+03
+1.100000000000000000e+03
+1.200000000000000000e+03
+1.300000000000000000e+03
+1.400000000000000000e+03
+1.500000000000000000e+03
+1.600000000000000000e+03
+1.700000000000000000e+03
+1.800000000000000000e+03
+1.900000000000000000e+03
+2.000000000000000000e+03
+2.100000000000000000e+03
+2.200000000000000000e+03
+2.300000000000000000e+03
+2.400000000000000000e+03
+2.500000000000000000e+03
+2.600000000000000000e+03
+2.700000000000000000e+03
+2.800000000000000000e+03
+2.900000000000000000e+03
+3.000000000000000000e+03
