### 2024-02-07 ebb425f
find closest temperature, pressure, wl ind
