### 2024-02-09 8bbadad
test linalg error for bad cube
