### 2024-02-09 f91a886
check more PCA error
