### 2024-02-09 b69ba79
update the chunking type situation
