+++ src/cortecs/tests/test_chunking.py
@@ -53,4 +53,30 @@ class TestIntegration(unittest.TestCase):
             opac_obj_ref.wl, np.concatenate((opac_obj0.wl, opac_obj1.wl))
         )
 
+    def test_vals_of_each_created_file(self):
+        """
+        Test that the opacity values together are concatenated!
+        """
+        # clean up the files already made!!
+        os.remove(self.first_file)
+        os.remove(self.second_file)
+
+        chunk_wavelengths(self.opacity_file, wav_per_chunk=2)
+
+        # now get the wavelengths of each file
+        opac_obj_ref = Opac(self.opacity_file, loader="exotransmit")
+        opac_obj0 = Opac(
+            self.first_file,
+            loader="exotransmit",
+        )
+        opac_obj1 = Opac(
+            self.second_file,
+            loader="exotransmit",
+        )
+        # pdb.set_trace()
+        np.testing.assert_array_equal(
+            opac_obj_ref.cross_section,
+            np.concatenate((opac_obj0.cross_section, opac_obj1.cross_section)),
+        )
+
         # self.assertEqual(opac_obj_ref.wl), len(opac_obj0.wl) + len(opac_obj1.wl))
