+++ src/cortecs/tests/test_chunking.py
@@ -4,7 +4,7 @@ just do a few tests for chunking...
 import unittest
 import os
 import numpy as np
-from cortecs.opac.chunking import chunk_wavelengths
+from cortecs.opac.chunking import chunk_wavelengths, add_overlap
 from cortecs.opac.opac import Opac
 
 
@@ -14,6 +14,7 @@ class TestIntegration(unittest.TestCase):
     second_file = (
         os.path.abspath(".") + "/src/cortecs/tests/" + "opacCH4_narrow_wl1.dat"
     )
+    file_base = os.path.abspath(".") + "/src/cortecs/tests/" + "opacCH4_narrow_wl"
 
     def test_chunking_two_files(self):
         """
@@ -77,6 +78,44 @@ class TestIntegration(unittest.TestCase):
         np.testing.assert_array_equal(
             opac_obj_ref.cross_section,
             np.concatenate((opac_obj0.cross_section, opac_obj1.cross_section)),
+        )
+
+        # self.assertEqual(opac_obj_ref.wl), len(opac_obj0.wl) + len(opac_obj1.wl))
+
+    def test_add_overlap_wl_increase_or_same(self):
+        """
+        Test that the overlap is added to the end of the first file.
+        """
+        # clean up the files already made!!
+        # os.remove(self.first_file)
+        # os.remove(self.second_file)
+        #
+        # chunk_wavelengths(self.opacity_file, wav_per_chunk=2, add_overlap=True)
+        opac_obj0_orig = Opac(
+            self.first_file,
+            loader="exotransmit",
+        )
+        opac_obj1_orig = Opac(
+            self.second_file,
+            loader="exotransmit",
+        )
+
+        add_overlap(self.file_base)
+
+        # now get the wavelengths of each file
+        opac_obj_ref = Opac(self.opacity_file, loader="exotransmit")
+        opac_obj0 = Opac(
+            self.first_file,
+            loader="exotransmit",
+        )
+        opac_obj1 = Opac(
+            self.second_file,
+            loader="exotransmit",
+        )
+        # pdb.set_trace()
+        self.assertTrue(
+            len(opac_obj0.wl) >= len(opac_obj0_orig.wl)
+            and len(opac_obj1.wl) >= len(opac_obj1_orig.wl)
         )
 
         # self.assertEqual(opac_obj_ref.wl), len(opac_obj0.wl) + len(opac_obj1.wl))
