### 2024-02-09 5febf6c
add overlap test with a single overlapped point
