### 2024-02-09 #6104 @arjunsavel
Thanks all! I'm cleaning up my next round of responses now.

@pscicluna, I've made a new branch over at https://github.com/arjunsavel/cortecs/tree/diff_filetype. I've output some of the [quickstart data](https://github.com/arjunsavel/cortecs/tree/main/docs/pages) into text files there for you to check out. Please let me know if that works for you — we can definitely iterate in an issue in the cortecs repo. Thanks again for following up on this!
