### 2024-02-09 #6104 @arjunsavel
also, @ivastar I'm thinking of submitting this paper as an arxiv preprint sometime soon — should I just ping the editorialbot?
