### 2024-02-09 f6e65f8
pass the fitting args to eval!
