+++ src/cortecs/eval/eval_pca.py
@@ -7,7 +7,7 @@ import numpy as np
 
 
 @jax.jit
-def eval_pca_ind_wav(temperature_ind, pressure_ind, vectors, pca_coeffs):
+def eval_pca_ind_wav(first_ind, second_ind, vectors, pca_coeffs):
     """
     Evaluates the PCA fit at a given temperature and pressure.
 
@@ -17,10 +17,10 @@ def eval_pca_ind_wav(temperature_ind, pressure_ind, vectors, pca_coeffs):
 
     Inputs
     ------
-    temperature_ind: int
-        The index of the temperature to evaluate at.
-    pressure_ind: int
-        The index of the pressure to evaluate at.
+    first_ind: int
+        The index of the first axis quantity (default temperature) to evaluate at.
+    second_ind: int
+        The index of the second axis quantity (default pressure) to evaluate at.
     vectors: array
         The PCA vectors.
     pca_coeffs: array
@@ -34,9 +34,7 @@ def eval_pca_ind_wav(temperature_ind, pressure_ind, vectors, pca_coeffs):
     xsec_val = 0.0
     n_components = vectors.shape[1]
     for component in range(n_components):
-        xsec_val += (
-            vectors[temperature_ind, component] * pca_coeffs[component, pressure_ind]
-        )
+        xsec_val += vectors[first_ind, component] * pca_coeffs[component, second_ind]
     return xsec_val
 
 
@@ -66,8 +64,6 @@ def eval_pca(
 
     """
     # find the nearest temperature, pressure, and wavelength indices.
-    print("this is my fit axis")
-    print(fit_axis)
 
     temperature_ind = np.argmin(np.abs(T - temperature))
     pressure_ind = np.argmin(np.abs(P - pressure))
@@ -77,12 +73,14 @@ def eval_pca(
     pca_coeffs = pca_coeffs_all_wl[wavelength_ind, :, :]
 
     # todo: figure out how to order the pressure and temperature inds!
+    # pdb.set_trace()
     if fit_axis == "pressure":
-        first_arg = temperature_ind
-        second_arg = pressure_ind
-    elif fit_axis == "temperature":
         first_arg = pressure_ind
         second_arg = temperature_ind
+    elif fit_axis == "temperature":
+        first_arg = temperature_ind
+        second_arg = pressure_ind
+
     elif fit_axis == "best":
         T_length = len(T)
         P_length = len(P)
@@ -94,5 +92,6 @@ def eval_pca(
         else:
             first_arg = pressure_ind
             second_arg = temperature_ind
-
+    # print("first_arg, second_arg", first_arg, second_arg)
+    # print("shapes:", pca_vectors.shape, pca_coeffs.shape)
     return eval_pca_ind_wav(first_arg, second_arg, pca_vectors, pca_coeffs)
