### 2024-02-09 aaeae1a
reduce error tol by OOM
