+++ src/cortecs/tests/test_integration.py
@@ -208,7 +208,7 @@ class TestIntegration(unittest.TestCase):
             }
         )
 
-    def test_pca_different_axis(self):
+    def test_pca_temperature_axis(self):
         """
         test that the PCA fit works well enough even when fit along a different axis.
         """
@@ -221,7 +221,33 @@ class TestIntegration(unittest.TestCase):
             self.cross_sec_filename, loader="platon", load_kwargs=load_kwargs
         )
         # fit axis is temperature now
-        fitter = Fitter(opac_obj, method="pca", wav_ind=-2, nc=3, fit_axis="pressure")
+        fitter = Fitter(
+            opac_obj, method="pca", wav_ind=-2, nc=3, fit_axis="temperature"
+        )
+        fitter.fit()
+        # run the metrics to see what the median absolute deviation is
+        vals, orig_vals, abs_diffs, percent_diffs = calc_metrics(
+            fitter, tp_undersample_factor=1, wl_under_sample_factor=8, plot=False
+        )
+
+        # check that the median absolute deviation is less than 10%
+        median_err = np.median(np.abs(abs_diffs))
+        self.assertTrue(median_err < 1)  # a reasonable fit
+
+    def test_pca_best_axis(self):
+        """
+        test that the PCA fit works well enough even when fit along a different axis.
+        """
+        load_kwargs = {
+            "T_filename": self.T_filename,
+            "P_filename": self.P_filename,
+            "wl_filename": self.wl_filename,
+        }
+        opac_obj = Opac(
+            self.cross_sec_filename, loader="platon", load_kwargs=load_kwargs
+        )
+        # fit axis is temperature now
+        fitter = Fitter(opac_obj, method="pca", wav_ind=-2, nc=3, fit_axis="best")
         fitter.fit()
         # run the metrics to see what the median absolute deviation is
         vals, orig_vals, abs_diffs, percent_diffs = calc_metrics(
