### 2024-02-10 5df8012
fix a little paper typesetting
