### 2024-02-14 4f05ca4
add conda badge [skip ci]
