### 2024-02-14 790efaf
add conda install back into manuscript
