### 2024-02-14 7633cf4
put preprint badge next to joss badge [skip ci]
