### 2024-02-14 a11f3fe
add ipykernel to doc requirements. this will let docs execute
