+++ docs/doc_requirements.txt
@@ -11,3 +11,4 @@ tensorflow
 h5py
 mock
 matplotlib
+ipykernel
