+++ docs/conf.py
@@ -47,10 +47,10 @@ import sys
 # for mod_name in MOCK_MODULES:
 #     sys.modules[mod_name] = mock.Mock()
 
-sys.path.insert(0, os.path.abspath("../.."))
-sys.path.insert(0, os.path.abspath(".."))
+# sys.path.insert(0, os.path.abspath("../.."))
+# sys.path.insert(0, os.path.abspath(".."))
+# sys.path.insert(0, os.path.abspath("../src"))
 sys.path.insert(0, os.path.abspath("../src"))
-sys.path.insert(0, os.path.abspath("../src/cortecs"))
 
 
 # -- General configuration ------------------------------------------------
