+++ docs/conf.py
@@ -50,7 +50,7 @@ import sys
 # sys.path.insert(0, os.path.abspath("../.."))
 # sys.path.insert(0, os.path.abspath(".."))
 # sys.path.insert(0, os.path.abspath("../src"))
-sys.path.insert(0, os.path.abspath("../src"))
+sys.path.insert(0, os.path.abspath("../../src"))
 
 
 # -- General configuration ------------------------------------------------
