+++ src/cortecs/opac/interpolate_cia.py
@@ -18,7 +18,7 @@ import os
 from glob import glob
 
 import numpy as np
-from tqdm import tqdm
+from tqdm.autonotebook import tqdm
 
 from cortecs.opac.chunking import *
 from cortecs.opac.io import *
