+++ src/cortecs/opt/optimize_pca.py
@@ -5,7 +5,7 @@ for computing eigenvectors.
 import math
 
 import numpy as np
-from tqdm import tqdm
+from tqdm.autonotebook import tqdm
 
 from cortecs.fit.fit import Fitter
 from cortecs.fit.fit_neural_net import *
