### 2024-02-15 a4598ec
just tqdm notebook for fit
