### 2024-02-15 #6104 @arjunsavel
Thank you again to the reviewers for such informative and thoughtful responses! We’ve addressed them below (with the minor exception of the readthedocs — currently wrestling with progress bars printing across lines).

Also, we’ve added back in that the project is available on conda, because it is now conda-installable: https://github.com/conda-forge/cortecs-feedstock

**Response to @doriannblain**

> The first one is that I still cannot reproduce the factor 13 compression claimed in the paper, with PCA compression and nc = 3. I however did not try with the referenced opacities (which, as the authors state, have 18 pressure points instead of 13). Could the authors provide a link to the opacities used?

Yes, the opacities are located here: https://www.dropbox.com/sh/0cxfolfmrs8ip37/AACV3dp5X3tJ-YpXaxkPtWE6a/RETRIEVAL/ABSCOEFF?dl=0&subfolder_nav_tracking=1

> I think a more "theoretical" explanation on how the compression factor is achieved is necessary. For example, something along the line of "instead of having n pressure points, we now have (n/~13 =) 3 (?) PCA coefficients [...]".

We include a statement to this effect at the beginning of the quickstart notebook now: “Compressing with PCA essentially allows us to swap an NTEMP x NPRESSURE x NWAV array for an NTEMP x NCOMPONENT x NWAV array, achieving approximately NPRESSURE/NCOMPONENT compression. “

> For the pRT opacities, this is the code I needed, I added suggestions and questions in comments with # :

Thank you so much for trying this out on a new opacity format! This is a very instructive example. I've added these as io options. I could also add a tutorial showing how to add a new opacity source, which would expose some more functionality to users.

> I thank the authors for making the tensorflow and keras modules optional. However, these modules are still imported with cortecs.fit.fit (required to use the Fitter object), whether or not they have been installed. I suggest using a try/except ModuleNotFoundError block in the cortecs.fit.fit_neural_net submodule, or adding intermediate functions in the Fitter object that would import the required submodule functions as needed, e.g.:

We have wrapped this as suggested.

> I appreciate the changes made in the "Test the installation" section. However, python3 -m unittest discover src still does not run any test. I see that there are tests implemented, but this is what I obtain:

Ah, I misunderstood — thank you for clarifying. I’ve updated the tests command (python -m unittest discover src/cortecs/tests), and that should work now.

> I think the fact that the fitted T-P dependency of an opacity at the chosen wavelength (wav_ind) is assumed for all the opacity's wavelengths should be highlighted even more. I suggest adding a step where you explain how and why you chose wav_ind=-2.

We have highlighted as such: “Again, the temperature--pressure dependence at all wavelengths is fitted as a linear combination of the eigenvectors calculated at `wav_ind`.”

> Related to the above, I think your optimize_pca function can find the best (or a "good enough") wavelength to perform the fit. This is very useful, so maybe this should be highlighted here as well.

We have highlighted as such: “See the [optimizing fits notebook](https://cortecs.readthedocs.io/en/latest/pages/optimizing) for more information on how to choose the best `wav_ind`.”

> I noted several inconsistencies between the text and the code (e.g. "We pass nc=3", then in cell 7, nc=5). This should be fixed.

We fix these inconsistencies: “We pass nc=5”, and “wav_ind=-2”

> I appreciate the extended discussion on nc, but I think it should be extended even more. What happens if one increase/decrease nc? What are the benefits/trade-offs? This should be discussed.

We add the following phrase: “Increasing `nc` can provide a better fit, but doing so will take up more RAM. Decreasing `nc` may yield a worse opacity fit, but it will consume less memory.”

> "This is the wavelength at which the eigenvectors are found". Do you mean "calculated" instead of "found"?

Yes, we change to “calculated.”

> It is not clear in the example what the eval_pca arguments are and what they do. You could explicit them in the cell and describe them in the text.

Done.

> Currently, the save function saves the "fitting parameters" in 2 separate npy files. What about saving them in one npy file, or one npz file with numpy.savez_compressed, or even in a HDF file, since you have h5py as requirement?

Done! Now using savez_compressed.

> I think the single point error evaluation (cells 14--18) and related comments could be removed entirely to only show the results from the calc_metrics function, which provide a more global view of the error.

Done.

> The accuracy map in cell 6 that I obtain is different from the one in the documentation. I suspect it has not been updated.

We’ve rerun this and enforced seeds for reproducibility. Thank you!

> The function save_neural_network does not exist. It should be save_neural_net (although in my opinion, the first name is the better one).

We’ve changed to “save_neural_net” for now (will change in future)

> The "Fitting with polynomials" example has not been updated, as I obtain a different value at the end of cell 31.

We’ve fixed this.

> Same with the "Fitting with polynomials" example. Also, the all_wl argument does not exist.

We removed the “all_wl” argument.

> I am glad that this is implemented. However, I think there should be no sampling of T and P by default. In the "Quickstart" example, sampling can be used but they should be explicited.

Done!

> Also, the function requires the attribute species_weight, which is not in e.g. loader_base.

Yes, the function assumes that the opac object has been loaded in with the species_weight attribute assigned. I’ve added a try / except clause in the calc_metrics function. thoughts?

> l. 35: you could mention that in many cases, not all wavelengths need to be loaded.

We have added: “In many cases, not all wavelengths need to be loaded, e.g. if the user is down-sampling resolution.”

> l. 49: I suggest rephrasing the sentence to make it clearer that "reasonable" comes from the benchmark, e.g. " Additionally, our benchmark only shows that the amounts of error from our compression technique is reasonable in the spectra of exoplanet atmospheres at pressures greater than a microbar for a single composition".

Done.

> l. 56 and 57: you should add a reference to support this claim.

Done (and relaxed to “often modeled in hydrostatic equilibrium”

> l. 63: I acknowledge that this is a hypothetical statement, but are you aware of an opacity with a sharp temperature-pressure dependence? You could add it as an example.

We’ve added a nod to the temperature-sensitive Lyman-alpha transitions.

> I find the transparent lines hardly visible. Maybe zoom-in around 1.8 micron?

We have zoomed accordingly.

> The units on the top panel and the units of the corresponding residuals on the bottom panel are not the same. Why?

Unit error. Thanks for catching!

> Fix the "$" left in the captions.

Done.

> About the 1e27 m column, I agree that you need an unrealistically long column, but are you sure of this value?

I did an order of magnitude estimate — should I link the calculation in the paper? I solved ds =  tau /  alpha, where tau=1 for the optically thick scenario, alpha = kappa * mass density, mass density is derived from ideal gas, and kappa is set to 1e-33 cm^2/g (equivalent to the cross-section of sigma ~ 5e-60 m^2). I also changed my value to 1e34 —my calculation was expressed in jupiter radii.

> l. 112: as said above, I still cannot get the 13 compression factor. There is also a residual draft comment ("cite them") left.

Removed the comment.

> l. 126: could you be more precise? An order of magnitude (e.g. a few minutes of differences for a run of several hours) would be enough.

We have clarified: “(about 5 hours of extra runtime on a roughly 2 day-long calculation)”

> Fig. 2: I thank the authors for fixing the gold and teal labels, but now they read identical values, while it is clear that the posteriors have some differences.

Copied the wrong labels — thanks for catching this!

> Table 1: I thank the authors for adding this table, but the decompression time for PCA (230 s) does not seem to match the value given at l. 120 (789 ns). Maybe in the first case the time is to obtain the opacities at all wavelengths, while in the other it is at one wavelength only? This should be mentioned.

Thank you for catching this — you are correct. We have clarified that the ns estimate is “Accessing/evaluating a single opacity”. And for the table, “The timing is over the entire wavelength range.”

> Thanks for the clarification. In that case, I suggest adding a comment in the code that this value was chosen to follow the PLATON code.

The comment now reads: “# set a floor to the opacities, as per the floor used in the PLATON code's opacity files.”

> Following the "Quickstart" example, I find instead that pca_coeffs is of shape (N_WAVE, N_PC, N_TEMP) (i.e. 4616, nc+1, 30), while vectors is of shape (N_PRESS, N_PC) (i.e. 13, nc+1). If that is not expected, then I suggest automatizing the fitted axe in favour of the second larger one (to prevent fitting over wavelengths).

Ah, great catch — thank you! We’ve automated this now.

> Indeed, I apologize. But in that case, you should prevent an error to be raised when the exact temperature/pressure needed is not passed. The function isclose is not sufficient, the code should instead fetch the nearest value. In the future, you could interpolate the opacities in that case instead.

Great point. We now eschew isclose for finding the nearest value.

> Yes indeed. Maybe at the beginning of "Quickstart" to have an overview of the possibilities right away?

Done!

> I think if you plan to discuss about Optimizer in this submission, then you should indeed implement it for the polynomial method. Otherwise, this can be the object of a future update.

Sure, we can have it be the object of a future update. It’s not listed in the paper, I believe.

**Response to @pscicluna**

> The description of potential failure modes is very helpful. The authors' suggestion of a figure showing it would be greatly appreciated! Seeing how it fails will help users diagnose their own issues. Adding this to the documentation as well would be great.

We have added a new Fig. 1 showing the optical depth and e^-tau as a function of grid size.

> The idea behind Figure 1 is definitely what I had in mind. However, I find it a bit difficult to read - the opacity spectra are so rich that it is hard to see any details. It is also unclear from the figure which compression method was used. It might be helpful to enlarge the figure and/or zoom in on narrow wavelength ranges of the spectra. I realise that the residual plot shows that the differences are generally small except in extreme cases, but seeing the details would be much more illustrative of the power of cortecs than the current plot it. Ideally, the results of multiple different compression methods would be shown on the same data, to give a sense of the trade-offs between them.

I’ve zoomed in on this region per the other reviewer’s suggestion. I’ve also made the line widths thinner and made the plot taller. Is this helpful?

> The improved API docs are greatly appreciated - however, it appears that these changes have not yet synced to RTD. I assume this is because no new version has been tagged yet, in which case it's not a big deal.

Thank you for catching this! We hadn’t included matplotlib in a few doc requirements, so a few modules could not be loaded. We have changed this, and the API should now [reflect this accordingly](https://cortecs.readthedocs.io/en/latest/pages/api/#optimize).

> I still find the description of some details a bit lacking. For example, the optimisation is incomplete. I see from the code that it is performing a grid-search to find the hyperparameters with the lowest MSE, but it would be helpful to have this described in the documentation. It's also not clear to me from the code how the maximum size of the compressed data is considered in the optimisation - the code to me looks like it is just performing a grid search over the hyperparameters, and not considering the size of the compressed data.

I’ve added this to the documentation: “This process essentially performs a grid search to find the hyperparameters with the lowest mean-squared error.” I’ve also removed the size of the compressed data as a consideration, as this was left over from an earlier version of the code.

> I noticed that the package requires that devs update the version number manually. This can be easy to forget, and result in a mismatch between tags and the version number in the code. It may be worth considering using setuptools-git-versioning and making the version a dynamic value in pyproject.toml. This ensures that the git tag (or even commit hash) is the ultimate arbiter of version numbers, and can help prevent mismatches. Since this would also involve updating the package from setup.py to pyproject.toml, it may be worth either waiting until the next major release, or trying to use python-versioneer instead of setuptools-git-versioning.

This is a great point. We’ll defer this until our next major release, so that we don’t have to greatly update the packaging at this point.

> I'm still facing issues with loading the data for the quickstart tutorial. This may be related to my environment, but I can't be sure.
> Running on Colab with numpy==1.23.5 I receive the following error:
	Interesting! Thank you for investigating this. These are the PLATON files, so I don’t want to go *too* far into redoing the format entirely. That being said, I have .txt files in the a new branch: https://github.com/arjunsavel/cortecs/tree/diff_filetype
