### 2024-02-15 b16bd92
remove pandoc version constraint
