+++ docs/doc_requirements.txt
@@ -12,5 +12,4 @@ h5py
 mock
 matplotlib
 ipykernel
-pandoc>2.14.2
 Sphinx<7.2.6.
