### 2024-02-15 815dddf
pin sphinx even lower, and lower nbsphinx
