+++ docs/doc_requirements.txt
@@ -3,7 +3,7 @@ jaxlib
 jax
 tqdm
 markupsafe==2.0.1
-nbsphinx
+nbsphinx<0.9.3
 sphinx_book_theme
 pandas
 keras
@@ -12,4 +12,4 @@ h5py
 mock
 matplotlib
 ipykernel
-Sphinx<7.1.6
+Sphinx<7.1
