+++ docs/doc_requirements.txt
@@ -12,4 +12,4 @@ h5py
 mock
 matplotlib
 ipykernel
-Sphinx==5.3.1
+Sphinx==5.3.0
