### 2024-02-27 5ae5f75
prevent test_fit warning
