### 2024-02-27 b2268ff
add opac kwargs
