### 2024-03-01 3666941
remove bad toy model plot
