### 2024-03-01 #6104 @arjunsavel
Thanks again to @doriannblain for such careful attention to our work! Responses below.

> I thank the author for providing the opacities they used in their paper; however, I still could not manage to reach the 13 compression factor reported: I obtained a factor 10 instead (which is still respectable). I might have done something wrong though. Below is the code I used.

Apologies for miscommunicating w.r.t. The number of PCA components. I indeed had NPC=2 for my calculations, and in the previous file format that led to sizes of 161.2 MB. My OS is also reporting the size of xsecarrCO_HITEMP_HELIOS_samp_3800_10000_R500000.h5 as 2.1 GB.

> Thanks for adding the table. However I'm surprised that in your article, the Polynomials compression has the lowest deviation, while here it has by far the largest. Does that mean that the accuracy of the different methods heavily depends on the dataset? If so, that must be highlighted.

We have added this in italics: *Note that the accuracy of the different methods heavily depends on the dataset.*

> About NCOMPONENT: if this is supposed to be the number of principal components, should not NCOMPONENT+1 instead of NCOMPONENT be used to represent the shape of pca_coeffs and vectors in the introduction?

Correct, thanks!

> Section "Fitting the opacity", it is stated that passing nc=5 will launch a fit with in "3" eigenvectors. I think it should be "5" instead.

Good catch, thanks.

> Instead of loading "test_pca.npz" twice, you could use:

Done, thanks for the suggestion!

> The optimizing notebook that is linked to serve as a guide for choosing wav_ind for PCA talks only about neural networks. You should at least add a work in progress section in that notebook for PCA optimization.

We briefly demonstrate PCA optimization now, as well.

> calc_metrics does not works with the "CHIMERA" opacities, as species_weight is still needed to run the function. Instead of using a try/except block, you could use an if block to handle the case where species_weight is not present in the Opac object. Alternatively, species_weight should be an argument common to all loader_base objects (and be set to 1 if not relevant).

Now using an `if` block!

> I think at the end of the notebook are leftover debug cells. These should be cleaned.

Done, thanks!

> In the fitting polynomial example, there are also some debugging cells left at the end of the notebook. These should be cleaned.

Done.

> In the neural network example, I still get results different from the ones in the documentation, despite the seed. I attached figures of what I obtain.

Interesting. I get results much closer to yours when I run the notebook on my local machine. There’s something interesting deep in the readthedocs OS setup…for now, I’ll mention that there’s some inherent discrepancies between OS, but that runs on an individual machine are reproducible.

> l.35: load_kwargs default value is a mutable, which can lead to unexpected behaviours because the default value can change during runtime. Instead, you should use None as default value and then if load_kwargs is None: \ load_kwargs = {}.

Done, thanks for the suggestion!

> load_kwargs could be named with the standard **kwargs instead, which would be even cleaner than the above. It depends on if you see Opac as an extension of loader_base or not though.

I’ve kept this as load_kwargs to differentiate between loader_kwargs, added based on the below suggestion.

> while I appreciate the addition of numerous useful attributes to the loader_base object, they cannot be set through an Opac object. There are multiple ways to fix this, but I think the quickest way would be to add a loader_kwargs argument to Opac.__init__, and pass it to _get_loader. Otherwise, I think that conceptually Opac should be a child of loader_base, but this would require a quite significant redisign of the code.

Added loader_kwargs argument.

> test_fit.py, l.99: this line is creating a warning during the tests, you could replace it by  np.ones((3, 3)) * np.nan.

Done!

> I think the documentation really lacks a detailed notebook on how to load the opacities (this could also be a "Quickstart" section as it would be rather small). You could simply list those next to their respective loader key. Right now one has to look into the source code to know that several opacities are (mostly) handled, which is not ideal. But as stated above, this can be the object of future updates.

We’ve added  this table in the quickstart.

> Fig. 1: I do not think this figure is necessary. The 3 points above detailing the limitations of cortecs are sufficient in my opinion. If you want to keep the figure, I have multiple remarks.

Sounds good — I've removed the figure.

> The label of the bottom panel should be log_10(absolute residual).

Changed to 'Absolute residuals of\n' + '$\log_{10}\sigma_{\lambda}$ (m$^2$)' (it’s the difference between the transparent and thick lines in the top plot)

> Mention explicitely in the caption that sigma_lambda is the opacity.

Done.

> Add the units of sigma_lambda in the caption.

Done!

> Mention that you cut-off your opacity at 1e-104, explaining the shape of the residuals in teal and dark red.

Done.

> Benchmark, 2nd paragraph: again, I find 194 MB instead of 154 MB.
See the response to the referee's first point.

> As mentioned above, the notebook does not talk about PCA optimizing, which is problematic. Otherwise, I think this part is clear enough.

We show PCA optimizing now, as well.

> That would be preferable. It is always a good thing to not have to find by yourself an equation if you do not know it. Otherwise, I agree with the 1e35 m value as put in the ArXiv version of the paper.

We’ve included the equation, and are now more precise about the value.
