### 2024-03-01 #6104 @arjunsavel
@ivastar @pscicluna i'm unable to reproduce the colab error with the pickle files, unfortunately.

I'm happy to move the quickstart away from pickle files and not use them internally. The issue is that some other tools (like PLATON) package their data in pickle files, which my package then interfaces with. I can chat with their developers to see if there's a workaround here!
