### 2024-03-11 d7b84a0
handle optional dep better
