### 2024-03-11 882a091
handle diff wavelenegth
