### 2024-03-11 arjunsavel/cortecs#32 @arjunsavel
calc metrics for different opacities?
calc_metrics may still not work as expected with non-PLATON opacities, because of the hardcoded conversion val = np.log10(val * species_weight * AMU * 1e-4).
