### 2024-03-11 f73edf3
change to 2 npc
