### 2024-03-11 #6104 @arjunsavel
Thanks for the helpful feedback, @doriannblain! I've addressed them below. I'll also regenerate the paper PDF in the next comment.

> The tensorflow optional dependency is not correctly handled. I suggest replacing the first lines of fit_neural_net.py with:
Thanks. We’ve implemented this better catch.

> The auto-detection of wavenumber sorting when loading an Opac object may not work properly in all cases. I suggest replacing l. 84 of io.py with: if np.all(np.diff(wl)) <= 0:. This is still quite dangerous as one can imagine e.g. a faulty opacity file where all wavenumber/wavelength are not correctly ordered. A more complex but more secure check would be ideal, but on the other hand this should work in almost all cases already.
Great point. We’ve updated this line as suggested.

> calc_metrics may still not work as expected with non-PLATON opacities, because of the hardcoded conversion val = np.log10(val * species_weight * AMU * 1e-4).
This seems like a bit of a refactor to do correctly, so we’ll punt to later work. We’ve made an issue: https://github.com/arjunsavel/cortecs/issues/32

> In the "Quickstart" notebook: you still report a 9% error in log10 abundance, but the error printed is 0.055%.
Done.

> About the compression factor, if you used nc=2, then in the paper it should read: "preserving 2 principal components". Also, please update the file sizes with those obtained with latest version.

Done.

> Also, please update the file sizes with those obtained with latest version.

Done.
