+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@
 # https://github.com/exoplanet-dev/exoplanet/blob/master/.pre-commit-config.yaml
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.4.0
+    rev: v4.5.0
     hooks:
       - id: trailing-whitespace
 
@@ -11,7 +11,7 @@ repos:
       - id: debug-statements
 
   - repo: https://github.com/pycqa/isort
-    rev: 5.12.0
+    rev: 5.13.2
     hooks:
       - id: isort
         name: isort (python)
@@ -29,19 +29,22 @@ repos:
         exclude: src/cortecs/tests
 
   - repo: https://github.com/psf/black
-    rev: 23.7.0
+    rev: 24.2.0
     hooks:
       - id: black
         language_version: python3 # Should be a command that runs python3.6+
+        additional_dependencies: [toml]
         exclude: src/cortecs/__init__.py
+        args: ["--target-version", "py38"]
 
   - repo: https://github.com/dfm/black_nbconvert
     rev: v0.4.0
     hooks:
       - id: black_nbconvert
+        additional_dependencies: [toml]
 
   - repo: https://github.com/adrienverge/yamllint.git
-    rev: v1.32.0
+    rev: v1.35.1
     hooks:
       - id: yamllint
         args: [-c=.yamllint]
