### 2024-03-12 122e45b
remove redundant init info
