+++ src/cortecs/__init__.py
@@ -1,9 +1,1 @@
 # -*- coding: utf-8 -*-
-
-__uri__ = "https://github.com/arjunsavel/cortecs"
-__author__ = "Arjun B. Savel, Megan Bedell, Eliza M.-R. Kempton"
-__email__ = "asavel@gmail.com"
-__license__ = "MIT"
-__version__ = "0.3.1"
-__release__ = "0.3.1"
-__description__ = "Compress opacity for radiative transfer"
