### 2024-03-12 1a77a01
add installation instructions to readme! [skip ci]
