+++ docs/conf.py
@@ -20,6 +20,10 @@ import os
 import re
 import sys
 
+from sphinx_pyproject import SphinxConfig
+
+config = SphinxConfig("../pyproject.toml", globalns=globals())
+
 # MOCK_MODULES = [
 #     "numpy",
 #     "scipy",
@@ -66,30 +70,6 @@ sys.path.insert(0, os.path.abspath("../../src"))
 META_PATH = os.path.join("..", "src", "cortecs", "__init__.py")
 HERE = os.path.abspath(os.path.dirname(__file__))
 
-
-def read(*parts):
-    """
-    Build an absolute path from *parts* and and return the contents of the
-    resulting file.  Assume UTF-8 encoding.
-    """
-    with codecs.open(os.path.join(HERE, *parts), "rb", "utf-8") as f:
-        return f.read()
-
-
-def find_meta(meta):
-    """
-    Extract __*meta*__ from META_FILE.
-    """
-    meta_match = re.search(
-        r"^__{meta}__ = ['\"]([^'\"]*)['\"]".format(meta=meta), META_FILE, re.M
-    )
-    if meta_match:
-        return meta_match.group(1)
-    raise RuntimeError("Unable to find __{meta}__ string.".format(meta=meta))
-
-
-META_FILE = read(META_PATH)
-
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.doctest",
@@ -132,9 +112,9 @@ author = "Arjun B. Savel, Megan Bedell, Eliza M.-R. Kempton"
 # built documents.
 #
 # The short X.Y version.
-version = find_meta("version")
-# The full version, including alpha/beta/rc tags.
-release = find_meta("release")
+# version = find_meta("version")
+# # The full version, including alpha/beta/rc tags.
+# release = find_meta("release")
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
