+++ docs/doc_requirements.txt
@@ -14,3 +14,4 @@ matplotlib
 nbconvert<7.14
 ipykernel
 Sphinx
+sphinx-pyproject
