### 2024-03-12 9b37c67
install own package in docs!!
