### 2024-03-12 0389b8e
use importlib to get cortecs version
