### 2024-03-12 823284e
actually install the things i know loaclly
