### 2024-03-12 3b0aa74
pin TF version so there is no memory overflow
