+++ pyproject.toml
@@ -42,7 +42,7 @@ dynamic = ["version"]
 
 
 [project.optional-dependencies]
-neural_networks = ["tensorflow<2.16",
+neural_networks = ["tensorflow<=2.15.0",
                 "keras"]
 
 [project.urls]
