### 2024-03-12 d20f091
update npz file in paper
