+++ .github/workflows/pythonpublish.yml
@@ -28,5 +28,5 @@ jobs:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
         run: |
-          python build
+          python -m build
           twine upload dist/*
