+++ .github/workflows/pythonpublish.yml
@@ -23,6 +23,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install setuptools wheel twine build
+          pip install -e .
       - name: Build and publish
         env:
           TWINE_USERNAME: __token__
