+++ pyproject.toml
@@ -39,6 +39,8 @@ dependencies = [
 ]
 dynamic = ["version"]
 
+[tool.setuptools_scm]
+
 
 
 [project.optional-dependencies]
