+++ src/cortecs/tests/test_integration.py
@@ -1,6 +1,7 @@
 """
 Test what happens when you link it all together.
 """
+
 import unittest
 from cortecs.fit.fit import *
 from cortecs.opac.opac import *
@@ -177,22 +178,7 @@ class TestIntegration(unittest.TestCase):
             self.cross_sec_filename, loader="platon", load_kwargs=load_kwargs
         )
         fitter = Fitter(opac_obj, method="neural_net")
-        # res = cortecs.fit.fit_neural_net.fit_neural_net(
-        #     fitter.opac.cross_section[:, :, -2],
-        #     fitter.opac.T,
-        #     fitter.opac.P,
-        #     None,
-        #     all_wl=False,
-        #     n_layers=3,
-        #     n_neurons=8,
-        #     activation="sigmoid",
-        #     learn_rate=0.04,
-        #     loss="mean_squared_error",
-        #     epochs=4000,
-        #     verbose=0,
-        #     sequential_model=None,
-        #     plot=False,
-        # )
+
         optimizer = Optimizer(fitter)
         max_size = 1.6
         max_evaluations = 8
