+++ src/cortecs/opac/io.py
@@ -3,6 +3,7 @@ Reads opacity data from various sources.
 
 author: @arjunsavel
 """
+
 import pickle
 
 import h5py
@@ -80,6 +81,7 @@ class loader_base(object):
 
         if self.wl_style == "wno":
             wl = 1e4 / wl
+            wl *= 1e-6  # now in meters
 
         if np.all(np.diff(wl)) <= 0:
             wl = wl[::-1]
@@ -353,22 +355,33 @@ class loader_exotransmit(loader_base):
 
         # read through all lines in the opacity file
         # skip through the header!
-        for x in tqdm(f1[2:], desc="reading wavelengths"):
-            # check if blank line
+        for x in tqdm(f1[2:], desc="reading exotransmit wavelengths"):
+            # # check if blank line
             if not x:
                 continue
             # check if a wavelength line
-            commad = x.replace(" ", ",")
-            if len(np.array([eval(commad)]).flatten()) == 1:
-                wavelengths += [eval(x[:-1])]
+            # commad = x.replace(" ", ",")
+            # if len(np.array([eval(commad)]).flatten()) == 1:
+            #     wavelengths.append(eval(x[:-1]))
+            # else:
+            #     # the first entry in each opacity line is the pressure
+            #     opacity_string = x.split()[1:]
+            #     opacity_vals = [eval(opacity) for opacity in opacity_string]
+            #     opacities.append(opacity_vals)
+            if not x.strip():
+                continue
+                # check if a wavelength line
+            line_values = x.split()
+            if len(line_values) == 1:
+                wavelengths.append(float(line_values[0]))
             else:
                 # the first entry in each opacity line is the pressure
-                opacity_string = x.split()[1:]
-                opacity_vals = np.array([eval(opacity) for opacity in opacity_string])
-                opacities += [opacity_vals]
+                opacity_vals = [float(opacity) for opacity in line_values[1:]]
+                opacities.append(opacity_vals)
 
         f.close()
         # pdb.set_trace()
+        # oh this isn't reshaped haha
         del f1
         try:
             return np.array(wavelengths), np.array(opacities)
@@ -394,10 +407,17 @@ class loader_exotransmit(loader_base):
             pressures
         """
 
-        wl, opacities = self.get_lams_and_opacities(filename)
+        wl, cross_section = self.get_lams_and_opacities(filename)
         T, P = self.get_t_p(filename)
+        P /= 1e5  # internally in bars
 
-        return wl, T, P, opacities
+        cross_section[np.less(cross_section, 1e-104)] = 1e-104
+
+        # and now make it log10
+        cross_section = np.log10(cross_section)
+        cross_section = cross_section.reshape(len(T), len(P), len(wl))
+
+        return wl, T, P, cross_section
 
 
 # todo: put these in the opac class?
@@ -516,10 +536,22 @@ def get_n_species(CIA_file, verbose=False):
 
 
 class writer_base(object):
-    def __init__(self):
+    def __init__(
+        self,
+        wl_key="wno",
+        T_key="T",
+        P_key="P",
+        cross_section_key="xsec",
+        wl_style="wno",
+    ):
         """
         nothing to do here
         """
+        self.wl_key = wl_key
+        self.T_key = T_key
+        self.P_key = P_key
+        self.cross_section_key = cross_section_key
+        self.wl_style = wl_style
         pass
 
 
@@ -542,7 +574,7 @@ class writer_exotransmit_cia(writer_base):
         """
 
         new_string = []
-        print('optimized time')
+        print("optimized time")
         # don't want to write temp and wav
         columns = list(opac.cross_section.columns)
         if "temp" in columns:
@@ -588,20 +620,20 @@ class writer_exotransmit_cia(writer_base):
[…]
