+++ src/cortecs/tests/test_chunking.py
@@ -1,6 +1,7 @@
 """
 just do a few tests for chunking...
 """
+
 import unittest
 import os
 import numpy as np
@@ -9,6 +10,8 @@ from cortecs.opac.opac import Opac
 
 
 class TestIntegration(unittest.TestCase):
+    load_kwargs = {"fullfile": False}
+
     opacity_file = os.path.abspath(".") + "/src/cortecs/tests/opacCH4_narrow_wl.dat"
     first_file = os.path.abspath(".") + "/src/cortecs/tests/" + "opacCH4_narrow_wl0.dat"
     second_file = (
@@ -43,12 +46,10 @@ class TestIntegration(unittest.TestCase):
         # now get the wavelengths of each file
         opac_obj_ref = Opac(self.opacity_file, loader="exotransmit")
         opac_obj0 = Opac(
-            self.first_file,
-            loader="exotransmit",
+            self.first_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         opac_obj1 = Opac(
-            self.second_file,
-            loader="exotransmit",
+            self.second_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         np.testing.assert_array_equal(
             opac_obj_ref.wl, np.concatenate((opac_obj0.wl, opac_obj1.wl))
@@ -67,12 +68,10 @@ class TestIntegration(unittest.TestCase):
         # now get the wavelengths of each file
         opac_obj_ref = Opac(self.opacity_file, loader="exotransmit")
         opac_obj0 = Opac(
-            self.first_file,
-            loader="exotransmit",
+            self.first_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         opac_obj1 = Opac(
-            self.second_file,
-            loader="exotransmit",
+            self.second_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         # pdb.set_trace()
         np.testing.assert_array_equal(
@@ -92,12 +91,10 @@ class TestIntegration(unittest.TestCase):
 
         chunk_wavelengths(self.opacity_file, wav_per_chunk=2)
         opac_obj0_orig = Opac(
-            self.first_file,
-            loader="exotransmit",
+            self.first_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         opac_obj1_orig = Opac(
-            self.second_file,
-            loader="exotransmit",
+            self.second_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
 
         add_overlap(self.file_base, v_max=0.0)
@@ -105,12 +102,10 @@ class TestIntegration(unittest.TestCase):
         # now get the wavelengths of each file
         opac_obj_ref = Opac(self.opacity_file, loader="exotransmit")
         opac_obj0 = Opac(
-            self.first_file,
-            loader="exotransmit",
+            self.first_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         opac_obj1 = Opac(
-            self.second_file,
-            loader="exotransmit",
+            self.second_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         # pdb.set_trace()
         self.assertTrue(
@@ -128,12 +123,10 @@ class TestIntegration(unittest.TestCase):
 
         chunk_wavelengths(self.opacity_file, wav_per_chunk=2)
         opac_obj0_orig = Opac(
-            self.first_file,
-            loader="exotransmit",
+            self.first_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
         opac_obj1_orig = Opac(
-            self.second_file,
-            loader="exotransmit",
+            self.second_file, loader="exotransmit", load_kwargs=self.load_kwargs
         )
 
[…]
