### 2024-03-21 87d5d2d
a few more opacs
