### 2024-04-16 #6104 @arjunsavel
@pscicluna just checking in! Were you able to access the data in the different format? Would the best path forward be to remove the pickle files in the quickstart and warn users that working with PLATON opacities might throw an error?
