### 2024-05-27 #6104 @ivastar
@arjunsavel and @pscicluna, pinging this issue again. I'm not sure if the example I tested above really addresses the issue you encountered. Could you please check and comment? The submission is really close and it would be great if we can wrap it up in the next couple of weeks. Thank you!
