### 2024-07-23 #6104 @ivastar
@arjunsavel while we wait for the reviewer to respond, I have moved on to some of the closing tasks. I reviewed the text of the paper and I have some minor editorial comments, listed below. Please let me know if you have questions. Also, please review the DOI issues above and correct them in the manuscript.

Line 10: "such opacity data for even a single gas can" --> "such opacity data, even for a single gas, can"
Line 12: "gases to consider" --> "gases to include"
Line 20: Can you please specify which spectrographs these citations are for? Maybe in the Instrument/Telescope format?
Line 49: "(see Benchmark)" --> "(see the Benchmark section below)"
Line 66: "may not perform optimally for" --> "may not perform optimally in"
Lines 75, 77: please remove the quotation marks around compression, fitting and lossy. If you want to underscore these words, I recommend using italics or bold instead.
Line 86: move "model" before the citation, also this citation is odd, can you edit it so it reads Chollet et al.?
Line 90: "neglected" --> "ignored"
Line 92: Maybe write out 10^35 m in more intuitive units? Like this is 3x10^19 parsecs!!!
Figure 1: Caption says 10^34 m, text 10^35m
Lines 92-96: this is a bit hard to read. Maybe something like:

This is because,
to become optically thick at a pressure of 1 bar and temperature of 1000 K, a column would need to be nearly $10^{35}$m long. Here we show a brief derivation of this. The length of the column, $ds$ is
$ds = \frac{\tau}{\alpha}$, where $\tau$ is the optical depth, and $\alpha$ is the absorption coefficient.
Setting $\tau = 1$, we have $ds = \frac{1}{\alpha}$. The absorption coefficient is the product of the opacity and the density of the gas: $\alpha$ = $\kappa_\lambda \rho$. Therefore,
$ds = \frac{1}{\kappa_\lambda \rho}$. The density of the gas $\rho$ is the pressure divided by the product of the temperature and the gas constant: $\rho = \frac{\rho}{k_BT\mu}$ for mean molecular weight $\mu$.
Which leads to the final equation for the column length: $ds = \frac{k_BT\mu}{\kappa_\lambda P}$ . For CO, the mean molecular weight is 28.01 g/mol.
Plugging in, we arrive at $ds \approx 10^{34}$m for $kappa_\lambda = 10^{-33} cm^2/g$, which is equivalent to roughly
$\sigma_\lambda = 10^{-60} m^2$.

Line 98: what is \sigma_\lambda?
Line 127: "as described in the previous paragraph" -- > as mentioned earlier in the paper?
