+++ paper/paper.bib
@@ -28,7 +28,7 @@
   pages={4884--4909},
   year={2020},
   publisher={Oxford University Press},
-  doi={doi.org/10.1093/mnras/staa548}
+  doi={10.1093/mnras/staa548}
 }
 
 @article{line:2021,
@@ -357,12 +357,13 @@ archivePrefix = {arXiv},
   year={2023},
   publisher={Oxford University Press}
 }
-@misc{chollet:2015,
+@misc{chollet2015keras,
   title={Keras},
-  author={Chollet, Fran\c{c}ois and others},
+  author={Chollet, Fran\c{c}ois},
   year={2015},
   howpublished={\url{https://keras.io}},
 }
+
 
 @article{polyansky:2018,
   title={ExoMol molecular line lists XXX: a complete high-accuracy line list for water},
