+++ paper/paper.md
@@ -29,9 +29,9 @@ bibliography: paper.bib
 
 The absorption and emission of light by exoplanet atmospheres encode details of atmospheric
 composition, temperature, and dynamics. Fundamentally, simulating these processes requires detailed knowledge
-of the opacity of gases within an atmosphere. When modeling broad wavelength ranges at high resolution, such opacity data for
-even a single gas can take up multiple gigabytes of system random-access memory (RAM). This aspect can be a limiting
-factor when considering the number of gases to consider in a simulation, the sampling strategy used for inference, or even the architecture of
+of the opacity of gases within an atmosphere. When modeling broad wavelength ranges at high resolution, such opacity data, for
+even a single gas, can take up multiple gigabytes of system random-access memory (RAM). This aspect can be a limiting
+factor when considering the number of gases to include in a simulation, the sampling strategy used for inference, or even the architecture of
 the system used for calculations. Here, we present `cortecs`, a Python tool for compressing
 opacity data. `cortecs` provides flexible methods for fitting the
 temperature, pressure, and wavelength dependencies of opacity data and for evaluating the opacity with accelerated,
@@ -39,7 +39,7 @@ GPU-friendly methods. The package is actively developed on GitHub (<https://gith
 available for download with `pip` and `conda`.
 
 # Statement of need
-Observations with the latest high-resolution spectrographs [e.g., @mace:2018; @seifahrt:2020; @pepe:2021]
+Observations with the latest high-resolution spectrographs [e.g., IGRINS / Gemini South, ESPRESSO / VLT, MAROON-X / Gemini North; @mace:2018; @seifahrt:2020; @pepe:2021]
 have motivated RAM-intensive simulations of exoplanet atmospheres at high spectral resolution.
 `cortecs` enables these simulations with more
 gases and on a broader range of computing architectures by compressing opacity data.
@@ -67,7 +67,7 @@ the temperature and pressure dependencies are generally smooth and similar acros
 This feature implies that the opacity data should be compressible without significant loss of
 accuracy at the spectrum level.
 
-While our benchmark case (see Benchmark) demonstrates the applicability of `cortecs` to high-resolution
+While our benchmark case (see the Benchmark section below) demonstrates the applicability of `cortecs` to high-resolution
 opacity functions of molecular gases, the package is general and the compression/decompression steps of the package can be applied to any opacity data in HDF5 format that has pressure and temperature dependence, such as the opacity of neutral atoms or ions. Our benchmark only
 shows, however, that the amounts of error from our compression technique is reasonable in the spectra of exoplanet atmospheres
 at pressures greater than a microbar for a single composition. This caveat is important to note for a few reasons:
@@ -80,7 +80,7 @@ atmospheres (in our case, 100) is small enough that the `cortecs` error is not l
 When modeling atmospheres in hydrostatic equilibrium, the final spectrum essentially maps to the altitude at which
 the gas becomes optically thick. If `cortecs`-compressed opacities were used to model an optically thin gas over
 large path lengths, however, then smaller opacities would be more important. `cortecs` tends to perform worse at
-modeling opacity functions that jump from very low to very high opacities, so it may not perform optimally for these
+modeling opacity functions that jump from very low to very high opacities, so it may not perform optimally in these
 optically thin scenarios.
 3. The program may perform poorly for opacity functions with sharp features in their temperature--pressure dependence
 [e.g., the Lyman series transitions of hydrogen, @kurucz2017including].
[…]
