### 2024-07-23 #6104 @arjunsavel
@ivastar thank you for your comments! I've addressed them and fixed the invalid DOI.

I don't think there are DOIs available for Keras and JAX. Should I just leave the bibtex entries as-is?
