### 2024-07-25 #6104 @ivastar
Regarding the missing DOIs, yes, I see that these are the recommended citations. Let's leave it at that.

Regarding the other comments, there are a couple of formatting issues still:

- On line 20-22, the braces should be round, not square.
- On Line 100 there is some equation formatting issue
- Figure 1: Caption says 10^34 m, text 10^35m, is this correct?
