### 2024-07-25 e82cfe0
fix keras citation
