+++ paper/paper.bib
@@ -357,7 +357,7 @@ archivePrefix = {arXiv},
   year={2023},
   publisher={Oxford University Press}
 }
-@misc{chollet2015keras,
+@misc{chollet:2015,
   title={Keras},
   author={Chollet, Fran\c{c}ois},
   year={2015},
