### 2024-08-01 #6104 @arjunsavel
@ivastar very glad to hear!

- version number: v1.0.0, link here: https://github.com/arjunsavel/cortecs/releases/tag/v1.0.0
- Zenodo link: https://zenodo.org/records/13137065
- Zenodo doi: https://doi.org/10.5281/zenodo.13137065
- authors, affiliations, license etc. checked.
