### 2024-08-02 #6104 @arjunsavel
@ivastar should I remove the duplicate entries to my .bib file? If so, should I also make a new package version / zenodo entry? Apologies for not catching this earlier!
