+++ paper/paper.bib
@@ -377,26 +377,3 @@ archivePrefix = {arXiv},
   publisher={Oxford University Press}
 }
 
-@article{barber:2014,
-  title={ExoMol line lists--III. An improved hot rotation-vibration line list for HCN and HNC},
-  author={Barber, RJ and Strange, JK and Hill, C and Polyansky, OL and Mellau, G Ch and Yurchenko, SN and Tennyson, Jonathan},
-  journal={Monthly Notices of the Royal Astronomical Society},
-  volume={437},
-  doi = {10.1093/mnras/stt2011},
-  number={2},
-  pages={1828--1835},
-  year={2014},
-  publisher={Oxford University Press}
-}
-
-@article{coles:2019,
-  title={ExoMol molecular line lists--XXXV. A rotation-vibration line list for hot ammonia},
-  author={Coles, Phillip A and Yurchenko, Sergei N and Tennyson, Jonathan},
-  journal={Monthly Notices of the Royal Astronomical Society},
-  volume={490},
-  number={4},
-  pages={4638--4647},
-  doi = {10.1093/mnras/stz2778},
-  year={2019},
-  publisher={Oxford University Press}
-}
