### 2024-08-04 #6104 @arjunsavel
@ivastar fixed:

version number: v1.0.1, link here: https://github.com/arjunsavel/cortecs/releases/tag/v1.0.1
Zenodo link: https://zenodo.org/records/13208759
Zenodo doi: https://doi.org/10.5281/zenodo.13208759
