### 2024-08-05 #6104 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
