### 2024-08-09 66296d3
Small edits to JOSS manuscript
