+++ paper/paper.bib
@@ -66,7 +66,7 @@
 @ARTICLE{august:2023,
        author = {{August}, Prune C. and {Bean}, Jacob L. and {Zhang}, Michael and {Lunine}, Jonathan and {Xue}, Qiao and {Line}, Michael and {Smith}, Peter C.~B.},
         title = "{Confirmation of Subsolar Metallicity for WASP-77Ab from JWST Thermal Emission Spectroscopy}",
-      journal = {\apjl},
+      journal = {Astrophysical Journal, Letters},
      keywords = {Exoplanets, Exoplanet atmospheres, James Webb Space Telescope, 498, 487, 2291, Astrophysics - Earth and Planetary Astrophysics},
          year = 2023,
         month = aug,
@@ -269,7 +269,7 @@ archivePrefix = {arXiv},
 @ARTICLE{prinoth:2023,
        author = {{Prinoth}, B. and {Hoeijmakers}, H.~J. and {Pelletier}, S. and {Kitzmann}, D. and {Morris}, B.~M. and {Seifahrt}, A. and {Kasper}, D. and {Korhonen}, H.~H. and {Burheim}, M. and {Bean}, J.~L. and {Benneke}, B. and {Borsato}, N.~W. and {Brady}, M. and {Grimm}, S.~L. and {Luque}, R. and {St{\"u}rmer}, J. and {Thorsbro}, B.},
         title = "{Time-resolved transmission spectroscopy of the ultra-hot Jupiter WASP-189 b}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {planets and satellites: atmospheres, planets and satellites: individual: WASP-189 b, planets and satellites: gaseous planets, techniques: spectroscopic, Astrophysics - Earth and Planetary Astrophysics},
          year = 2023,
         month = oct,
