### 2024-08-10 #6104 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
