### 2024-08-10 f6a3d87
Create CITATION.cff
